% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_mean_nets.R
\name{append_mean_nets}
\alias{append_mean_nets}
\alias{append_mean_nets.metanetwork}
\title{append mean networks}
\usage{
append_mean_nets(metanetwork)

\method{append_mean_nets}{metanetwork}(metanetwork)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}
}
\value{
object of class 'metanetwork' with mean networks appended per value of the covariable at all resolutions

\code{NULL}
}
\description{
Append mean networks per (discrete) value of covariable
}
\examples{
library(igraph)
library(metanetwork)


}
