% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySparrowBonferroni.R
\name{applySparrowBonferroni}
\alias{applySparrowBonferroni}
\title{Applies Bonferroni Correction to a Sparrow Network Object}
\usage{
applySparrowBonferroni(network)
}
\arguments{
\item{network}{A n x n upper triangular adjacency in the matrix class format.}
}
\value{
A n x n upper triangular adjacency in the matrix class format.
}
\description{
This function applies bonferroni correction to a sparrow network object.
}
