% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySparrowFDR.R
\name{applySparrowFDR}
\alias{applySparrowFDR}
\title{Applies FDR Correction to a Sparrow Network Object}
\usage{
applySparrowFDR(network)
}
\arguments{
\item{network}{A n x n upper triangular adjacency in the matrix class format.}
}
\value{
A n x n upper triangular adjacency in the matrix class format.
}
\description{
This function applies FDR correction to a sparrow network object.
}
