% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout.R
\name{attach_layout}
\alias{attach_layout}
\alias{attach_layout.metanetwork}
\title{compute and attach 'TL-tsne' layout}
\usage{
attach_layout(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)

\method{attach_layout}{metanetwork}(
  metanetwork,
  g = NULL,
  beta = 0.1,
  mode = "TL-tsne",
  TL_tsne.config = TL_tsne.default,
  res = NULL,
  group_layout.config = group_layout.default
)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{g}{character indicating the name of the network for which the 'TL-tsne' layout is computed,
default is 'metaweb'}

\item{beta}{the diffusion parameter of the diffusion kernel, a positive scalar controlling the
squeezing of the network}

\item{mode}{only 'TL-tsne' is supported for this function}

\item{TL_tsne.config}{configuration list for mode 'TL-tsne', default is TL_tsne.default}
}
\value{
'metanetwork' object with layout added as node attribute of the considered network

\code{NULL}
}
\description{
Method to compute 'TL-tsne' layout and save it as node attributes of the focal network. Each node of the focal network has an attribute \code{layout_beta_VALUE}.
If this function is run several times for a given beta value, repetitions of the layout algorithm will be stored as node attributes
}
\examples{
library(metanetwork)
library(igraph)
# on angola dataset (metaweb)
meta_angola = attach_layout(meta_angola,beta = 0.05)
V(meta_angola$metaweb)$layout_beta0.05
# on a local network
meta_angola = attach_layout(meta_angola,g = meta_angola$X1986,beta = 0.05)

# getting repetitions
meta_angola = attach_layout(meta_angola,beta = 0.05)
V(meta_angola$metaweb)$layout_beta0.05
V(meta_angola$metaweb)$layout_beta0.05_1

}
