% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3netWrapper.R
\name{c3netWrapper}
\alias{c3netWrapper}
\title{This function wraps c3net implementation}
\usage{
c3netWrapper(data, outputpath, c3net_alpha = 0.01, debug_save = FALSE)
}
\arguments{
\item{data}{Required. A Matrix containing gene expression values with samples
as row values and genes as column IDs.}

\item{outputpath}{Required. Path to save the resulting network}

\item{c3net_alpha}{Optional. The statistical significance threshold for c3net.
Default is 0.01, which is the default value for the `c3net()` function.}

\item{debug_save}{Optional. Whether to save the network matrix as an RDS
file, for debugging purposes.}
}
\value{
a vector of file names corresponding to the files generated by this
function. Currently the only file is: `<outputpath>/c3netNetwork.csv`.
}
\description{
This function wraps the c3net call on a gene expression matrix. The input
matrix is assumed to have samples as rows and genes as columns, so this
function takes the transpose of the matrix to be in the proper orientation
for c3net.
}
\references{
G. Altay, F. Emmert-Streib, "Inferring the conservative causal core of gene regulatory networks", BMC Systems Biology (2010) 4:132.
}
