% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHyperPval.R
\name{calcHyperPval}
\alias{calcHyperPval}
\title{Compute Hypergeometric PValue}
\usage{
calcHyperPval(geneSetToTest, moduleGenes, backgroundGenes)
}
\arguments{
\item{geneSetToTest}{Required. A character vector of gene IDs to test for an 
enrichment in `moduleGenes`.}

\item{moduleGenes}{Required. A character vector of gene IDs to test for being 
enriched with `geneSetToTest`.}

\item{backgroundGenes}{Required. A character vector of gene IDs to serves as 
the background gene set.}
}
\value{
numerical PValue
}
\description{
Computes a hypergeometric enrichment test of geneset `geneSetToTest` in a 
given module of genes `moduleGenes` given a total background set of genes 
`backgroundGenes`. Returns a P-Value of the enrichment test.
}
