% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.LocalModularity.R
\name{compute.LocalModularity}
\alias{compute.LocalModularity}
\title{Find Local Modularity (NQ)}
\usage{
compute.LocalModularity(adj, mod)
}
\arguments{
\item{adj}{Required. An n x n upper triangular adjacency in the matrix class
format.}

\item{mod}{Required. An n x 3 dimensional data frame with column names as
Gene.ID, moduleNumber, and moduleLabel.}
}
\value{
NQ = local modularity index.
}
\description{
This function finds local modularity (NQ) in an upper triangular adjacency matrix.
}
