% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Modularity.R
\name{compute.Modularity}
\alias{compute.Modularity}
\title{Find Global Modularity (Q)}
\usage{
compute.Modularity(adj, mod)
}
\arguments{
\item{adj}{Required. An n x n upper triangular adjacency in the matrix class
format.}

\item{mod}{Required. An n x 3 dimensional data frame with column names as
Gene.ID, moduleNumber, and moduleLabel.}
}
\value{
Q = modularity index.
}
\description{
This function calculates global modularity (Q) of a upper triangular
adjacency matrix using \code{igraph::modularity}.
}
