% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModularityDensity.R
\name{compute.ModularityDensity}
\alias{compute.ModularityDensity}
\title{Find Modularity Density (Qds)}
\usage{
compute.ModularityDensity(g, mod)
}
\arguments{
\item{g}{An \code{igraph} graph of the network}

\item{mod}{A named vector where names are genes and values are module membership}
}
\value{
Qds = module density.
}
\description{
This function calculates modularity density according to Botta and Genio 2016.
}
\references{
Federico Botta, Charo I. del Genio. Finding network communities
using modularity density. https://arxiv.org/abs/1612.07297
}
