% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarities.R
\name{compute_dissimilarities}
\alias{compute_dissimilarities}
\title{compute pairwise network dissimilarity indices}
\usage{
compute_dissimilarities(metanetwork, q = 1, res = NULL, ncores = 4)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{q}{viewpoint parameter controlling the weight given to abundant species/groups and links}
}
\value{
a list of \code{data.frame} containing node and link pairwise dissimilarities
}
\description{
Function to compute pairwise network dissimilarity indices based on Hill numbers following the method described in Ohlmann et al. 2019
}
\details{
XXXXXXXXXXXXXXXXXXXXXx
}
\examples{
library(metanetwork)
library(igraph)

#on angola dataset
data(meta_angola)
compute_dissimilarities(meta_angola,q = 1,ncores = 1)

#computing diversities only at Phylum level
compute_diversities(meta_angola,q = 1,res = "Phylum",ncores = 1)

}
\references{
Ohlmann, M., Miele, V., Dray, S., Chalmandrier, L., O'connor, L., & Thuiller, W. (2019). Diversity indices for ecological networks: a unifying framework using Hill numbers. Ecology letters, 22(4), 737-747.
}
