% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{compute network metrics}
\usage{
compute_metrics(metanetwork, res = NULL)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{res}{a vector containing the resolutions at which the metrics are computed}
}
\value{
a \code{data.frame}
}
\description{
Function to compute (some) network metrics on the metaweb and local networks
}
\details{
This function computes network metrics for the metaweb and local networks at the available resolutions. This function computes the following metrics:
\itemize{
\item weighted connectance: connectance weighted by abundances and link probability (see Ohlmann et al. 2019)
Denoting \eqn{p_q} abundance of node \eqn{q} and \eqn{\pi_{ql}} the interaction probability between nodes \eqn{q} and \eqn{l}, then the weighted connectance \eqn{C} is:
\deqn{C = \sum_{q,l} \pi_{ql}p_q p_l}
\item mean_TL and max_TL: mean and max trophic level of each network, using values computed from compute_TL() method.
\item mean_shortest_path_length: mean shortest path lenght of each network computed using igraph mean_distance() function.
}
}
\examples{
library(metanetwork)
library(igraph)

#on angola dataset
data("meta_angola")
compute_metrics(meta_angola)

#computing metrics only at Phylum level
compute_metrics(meta_angola,res = "Phylum")

}
\references{
\itemize{
\item Csardi, G., & Nepusz, T. (2006). The igraph software package for complex network research. InterJournal, complex systems, 1695(5), 1-9.
\item Ohlmann, M., Miele, V., Dray, S., Chalmandrier, L., O'connor, L., & Thuiller, W. (2019). Diversity indices for ecological networks: a unifying framework using Hill numbers. Ecology letters, 22(4), 737-747.
}
}
\seealso{
\code{\link[=compute_TL]{compute_TL()}},\code{\link[=compute_div]{compute_div()}}
}
