% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TL.R
\name{compute_TL}
\alias{compute_TL}
\alias{compute_TL.metanetwork}
\title{compute trophic levels}
\usage{
compute_TL(metanetwork)

\method{compute_TL}{metanetwork}(metanetwork)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}
}
\value{
object of class 'metanetwork' with a node attribude \code{TL}

\code{NULL}
}
\description{
Compute trophic levels using graph laplacian
}
\examples{
library(metanetwork)
library(igraph)

#on angola dataset
data(meta_angola)
meta_angola  = compute_TL(meta_angola)
V(meta_angola$metaweb)$TL

}
