% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBICcurve.R
\name{computeBICcurve}
\alias{computeBICcurve}
\title{This Function Computes a Network BIC Curve}
\usage{
computeBICcurve(network, exprData, maxEdges = NULL, exact = NULL)
}
\arguments{
\item{network}{A network object}

\item{exprData}{Expression data matrix}

\item{maxEdges}{Optional. (Default = NULL)}

\item{exact}{Optional.  (Default = NULL)}
}
\value{
A list object containing values of a sparse network, the best fit
minimum BIC, and
}
\description{
(?) TODO
}
