% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBICpath.R
\name{computeBICpath}
\alias{computeBICpath}
\title{Compute BIC path of a network's edges}
\usage{
computeBICpath(X, rankedEdges, startI = 1)
}
\arguments{
\item{X}{An expression matrix where rows are samples and columns are genes.}

\item{rankedEdges}{An edge list in the form of a matrix where column 1 is
gene one and column 2 is gene two}

\item{startI}{Optional. Which edge to start computing and keeping track of
BIC at in \code{rankedEdges} (Default = 1)}
}
\value{
A named list containing: "bic" = the vector of total BIC after each
  edge is added, "bicNeighborhood" = BIC for each individual gene and its
  neighborhood, "neighborhoods" = a list, one item for each gene, where each
  item is a vector of other genes in the neighborhood of that gene, "flag" =
  NULL, if no errors happen during calculation, or a 2-column data frame
  where column 1 is the name of the gene and column 2 is the edge number
  where errors occurred.
}
\description{
This function iteratively adds edges to a network from a list of ranked
edges, and computes the BIC of a linear model on each gene and its
neighborhood after adding each edge.
}
