% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.megena.R
\name{constructNetwork.megena}
\alias{constructNetwork.megena}
\title{Find Modules with Megena Clustering}
\usage{
constructNetwork.megena(
  data,
  n_cores = 1,
  log_file_path = ".",
  doPerm = 10,
  ...
)
}
\arguments{
\item{data}{A gene expression matrix where rows are samples and columns are
genes. The matrix will be transposed to work with MEGENA.}

\item{n_cores}{Optional. The number of cores/threads to use in parallel
computation. If n_cores = 1, MEGENA will not run in parallel.}

\item{log_file_path}{Optional. The folder path to where log files should be
stored. Log files capture any output during parallel execution. If omitted,
log files will be stored in the working directory.}

\item{doPerm}{Optional. Number of permutations for calculating FDRs for all
correlation pairs. (Default = 10)}

\item{...}{Optional. Additional arguments accepted by
\link[MEGENA:calculate.correlation]{MEGENA::calculate.correlation} or \link[MEGENA:calculate.PFN]{MEGENA::calculate.PFN}.}
}
\value{
An NxN matrix where N is the number of genes used in the network,
which may be less than the number of genes in \code{data}.
}
\description{
This function constructs a network adjacency matrix using MEGENA.
}
