% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.parallelRegression.R
\name{constructNetwork.parallelRegression}
\alias{constructNetwork.parallelRegression}
\title{Parallel Network Construction Wrapper}
\usage{
constructNetwork.parallelRegression(
  data,
  regressionFunction,
  log_file_path = ".",
  n_cores = 1,
  regulatorIndex = NULL,
  ...
)
}
\arguments{
\item{data}{Expression matrix to be used for network construction, which
should have samples as rows and genes as columns.}

\item{regressionFunction}{The name of the regression function to use. Current
options are: lassoIC, lassoCV, ridgeIC, ridgeCV, tigress, and vbsr}

\item{log_file_path}{Optional. The folder path to where log files should be
stored. Log files capture any output during parallel execution. If omitted,
log files will be stored in the working directory.}

\item{n_cores}{Optional. The number of parallel cores to use.}

\item{regulatorIndex}{Optional. A vector of numerical indexes into
\code{colnames(data)} for a subset of genes that should be used in the
network. All genes in \code{data} will be tested against this subset,
rather than against every gene. TODO should this stay?}

\item{...}{Optional. Additional arguments that are passed through to the
individual network algorithm.}
}
\value{
a named list, where each item is a network matrix as returned by the
  regression function. Regression functions may return multiple matrices or
  only one. If there was an error, the list will be empty.
}
\description{
Wrapper to run network detection in a parallel fashion, using a specified
regression function.
}
