% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.parmigene.R
\name{constructNetwork.parmigene}
\alias{constructNetwork.parmigene}
\title{Wrapper for MRNET, ARACNE, and CLR from \code{parmigene}}
\usage{
constructNetwork.parmigene(data, ...)
}
\arguments{
\item{data}{A gene expression matrix with rows as sample IDs and columns as
Gene or feature IDs. The matrix will be transposed to work with
\code{parmigene}.}

\item{...}{Optional. Additional arguments accepted by either
\code{parmigene::knnmi.all} or \code{parmigene::aracne.m}. The \code{clr}
and \code{mrnet} do not accept additional arguments.}
}
\value{
a named list of three matrices, where "aracne" is the output of
  \code{parmigene::aracne.m}, "clr" is the output of \code{parmigene::clr},
  and "mrnet" is the output of \code{parmigene::mrnet}.
}
\description{
This function uses \code{parmigene} to calculate network matrices using the
MRNET, ARACNE, and CLR algorithms.
}
\details{
A mutual information matrix is calculated with \code{parmigene::knnmi.all},
and then three network matrices are calculated from this mutual information
matrix using MRNET, ARACNE, and CLR. The results from all three algorithms
are returned as a named list.
}
