% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.wgcna.R
\name{constructNetwork.wgcna}
\alias{constructNetwork.wgcna}
\title{Wrapper for WGCNA}
\usage{
constructNetwork.wgcna(data, n_cores = 1, ...)
}
\arguments{
\item{data}{A gene expression matrix where rows are samples and columns are
genes.}

\item{n_cores}{Optional. How many cores/threads to allow when running WGCNA.
The default value of 1 means "do not thread".}

\item{...}{Optional. Additional arguments accepted by either
\code{WGCNA::pickSoftThreshold}, \code{WGCNA::adjacency}, or
\code{WGCNA::TOMsimilarity}.}
}
\value{
a named list of two network matrices, where "adjacency" is the output
  of \code{WGCNA::adjacency} and "TOM" is the output of
  \code{WGCNA::TOMsimilarity}.
}
\description{
Runs \code{WGCNA::adjacency} and \code{WGCNA::TOMsimilarity} over a gene
expression matrix.
}
