% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Compute a Correlation Matrix}
\usage{
correlation(data, path = NULL, method = "pearson", outputpath)
}
\arguments{
\item{data}{The input expression/numerical data matrix}

\item{path}{Optional. Currently not implemented in the function}

\item{method}{Optional. The type of correlation to be implemented. Options are
one of; c('pearson','spearman','kendall'). (Default = pearson)}

\item{outputpath}{The file path to save the correlation network to.}
}
\value{
Returns no value direct to R. This function saves the correlation network
as correlationNetwork.csv on the path specified by `outputpath`.
}
\description{
Compute a Correlation matrix from a an expression matrix and writes to a
user specified path. Correlation network is saved as correlationNetwork.csv on
the path specified by `outputpath`.
}
