% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceSelection.R
\name{covarianceSelection}
\alias{covarianceSelection}
\title{Covariance Selection with Graphical Lasso}
\usage{
covarianceSelection(S, rankedEdges)
}
\arguments{
\item{S}{Required. A symetric p-by-p covariance matrix.}

\item{rankedEdges}{Required. A list of ranked edges to be constrained by zero.}
}
\value{
A list with components.
\itemize{
   \item{`w` Estimated inverse covariance matrix.}
   \item{`loglik`	Value of maximized log-likelihodo+penalty.}
   \item{`errflag`	Memory allocation error flag: 0 means no error; !=0 means 
   memory allocation error - no output returned.}
   \item{`approx`	Value of input argument approx.}
   \item{`del` Change in parameter value at convergence.}
   \item{`niter`	Number of iterations of outer loop used by algorithm.}
}
}
\description{
Estimates a sparse inverse covariance matrix using a lasso (L1) penalty.
}
