% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceSelectionBisection.R
\name{covarianceSelectionBisection}
\alias{covarianceSelectionBisection}
\title{Covariance Selection with Bisection Optimization}
\usage{
covarianceSelectionBisection(
  S,
  rankedEdges,
  numberObservations,
  lowerBoundEdge,
  upperBoundEdge
)
}
\arguments{
\item{S}{Required. A symetric p-by-p covariance matrix.}

\item{rankedEdges}{Required. A list of ranked edges to be constrained by zero.}

\item{numberObservations}{Required. Number of observations used to calculate 
BIC estimates.}

\item{lowerBoundEdge}{Required. Numeric specifying the lower bound number of 
parameters (d) in BIC calculation: `BIC = -2 * loglikelihood + d * log(N)`}

\item{upperBoundEdge}{Required. Numeric specifying the upper bound number of 
parameters (d) in BIC calculation: `BIC = -2 * loglikelihood + d * log(N)`}
}
\value{
A list containg
\itemize{
\item{`w` Estimated inverse covariance matrix.}
   \item{`resMiddle`	The converged sparse inverse covariance matrix.}
   \item{`bicMiddle`	The converged BIC estimate.}
   \item{`middleEdge` The converged estimate of parameters.}
}
}
\description{
Selects an optimal covariance matrix through BIC convergence.
}
