% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceSelectionMBPath.R
\name{covarianceSelectionMBPath}
\alias{covarianceSelectionMBPath}
\title{Covariance based on Neighborhood Selection}
\usage{
covarianceSelectionMBPath(X, rankedEdges, numberObservations, startI = 1)
}
\arguments{
\item{X}{Required. An expression matrix}

\item{rankedEdges}{Required. An edge list in the form of a matrix where column
one is gene one and column 2 is gene two}

\item{numberObservations}{Required. The number of samples comprising the network (?)}

\item{startI}{Optional. Start at the first edge in `rankedEdges` (Default = 1)}
}
\value{
A list object of containing the BIC estimate, bicNeighborhood (?) , neighborhoods (?), flag (?)
}
\description{
This function defines the covariance neighborhood between gene-gene edges
in an expression matrix. TODO
}
