% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceSelectionPath.R
\name{covarianceSelectionPath}
\alias{covarianceSelectionPath}
\title{Covariance Selection with Path BIC Selection}
\usage{
covarianceSelectionPath(S, rankedEdges, numberObservations, n)
}
\arguments{
\item{S}{Required. An expression matrix}

\item{rankedEdges}{Required. An edge list in the form of a matrix where column
one is gene one and column 2 is gene two}

\item{numberObservations}{Required. The number of samples comprising the network (?)}

\item{n}{Optional. Start at the first edge in `rankedEdges` (Default = 1)}
}
\value{
A list containg
\itemize{
    \item{`bic` Estimated BIC.}
   \item{`bicmin` Lowest Estimated BIC.}
   \item{`w`	The covariance matrix.}
   \item{`wi` The inverse covariance matrix.}
}
}
\description{
Selects an optimal covariance matrix through Path BIC Selection.
}
