% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployNetworkSparsity.R
\name{deployNetworkSparsity}
\alias{deployNetworkSparsity}
\title{This function deploys arbitary sparsity on a list of networks}
\usage{
deployNetworkSparsity(network, sparsity)
}
\arguments{
\item{network}{Required. A matrix or data frame of gene expression}

\item{sparsity}{Required. Edges parameter to increase sparsity.}
}
\value{
a list of logical matrices of TRUE/FALSE where True indicates sparsity
}
\description{
This introduces sparsity into a list of data frames or matrices with the 
`arbitrarySparsity()` function with increasing frequency
as the `sparsity` parameter is specified
}
