% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{fastlm_bic}
\alias{fastlm_bic}
\title{Fast Linear Modeling BIC}
\usage{
fastlm_bic(y, x = NULL, correction = 1)
}
\arguments{
\item{y}{A numeric vector of model coefficients. If not set x
becomes a vector of integers the from 1 to length(y). (?)}

\item{x}{Optional. A numeric vector or matrix of model coefficients. If not set x
becomes a vector of integers the from 1 to length(y). (?)}

\item{correction}{A vector of correction factors, or a single numeric value.}
}
\value{
BIC estimate
}
\description{
This function runs \link{fastlm} and calculates the BIC score based on the
solution. TODO
}
