% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlmbic.R
\name{fastlmbic}
\alias{fastlmbic}
\title{Fast Linear Modeling BIC}
\usage{
fastlmbic(y, x = NULL, correction = 1)
}
\arguments{
\item{y}{A numeric vector of model coefficients. If not set x 
becomes a vector of integers the from 1 to length(y). (?)}

\item{x}{Optional. A numeric vector or matrix of model coefficients. If not set x 
becomes a vector of integers the from 1 to length(y). (?)}

\item{correction}{A vector of correction factors}
}
\value{
BIC estimate
}
\description{
This function deploys matrix operations to calculate a model BIC given a vector
of model coefficients.
}
