% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.CFinder.R
\name{findModules.CFinder}
\alias{findModules.CFinder}
\title{Find Modules with CFinder}
\usage{
findModules.CFinder(adj, path, nperm = 10, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{path}{File path location of CFinder.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
eneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
Function to get modules from network adjacency matrix using CFinder algorithm
}
