% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.GANXiS.once.R
\name{findModules.GANXiS.once}
\alias{findModules.GANXiS.once}
\title{Find Modules using GANXIS}
\usage{
findModules.GANXiS.once(adj, path, min.module.size)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function finds modules with GANXIS
}
