% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.GANXiS.R
\name{findModules.GANXiS}
\alias{findModules.GANXiS}
\title{Find Modules with GANXiS}
\usage{
findModules.GANXiS(adj, path, nperm = 10, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{path}{File path location of CFinder.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function to get modules from network adjacency matrix using GANXiS 
community detection algorithm v3.0.2.
}
