% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.hclust.R
\name{findModules.hclust}
\alias{findModules.hclust}
\title{Find Modules with Hierarchical Clustering}
\usage{
findModules.hclust(
  adj,
  aggloMethod = "ward",
  clustDistance = "euclidean",
  minModuleSize = 30
)
}
\arguments{
\item{adj}{Required. A n x n upper triangular adjacency in the matrix class format.}

\item{aggloMethod}{Optional. The agglomeration method to be used. This must be 
(an unambiguous abbreviation of) one of "single", "ward", "centroid" or "median".
(Default = "ward")}

\item{clustDistance}{Optional. The distance measure to be used. This must be 
one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". 
Any unambiguous substring can be given. (Default = "euclidean")}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function finds modules from Hierarchical Clustering `fastcluster::hclust.vector()`
}
