% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.infomap.once.R
\name{findModules.infomap.once}
\alias{findModules.infomap.once}
\title{Find Modules using graph adjacency}
\usage{
findModules.infomap.once(adj, min.module.size)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function finds modules with igraph::graph.adjacency()
}
