% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.infomap.R
\name{findModules.infomap}
\alias{findModules.infomap}
\title{Find Modules using graph adjacency}
\usage{
findModules.infomap(adj, nperm = 10, path, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{path}{File path location of CFinder.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function wraps an iterative process of finding modules with
igraph::graph.adjacency()
}
