% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.label_prop.R
\name{findModules.label_prop}
\alias{findModules.label_prop}
\title{Wraps Finding Modules with Graph Adjacency}
\usage{
findModules.label_prop(adj, nperm = 10, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
Deploys `igraph::graph.adjacency` iterativly to find module structure.
}
