% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.linkcommunities.R
\name{findModules.linkcommunities}
\alias{findModules.linkcommunities}
\title{Find Modules with Network Adjacency Matrix Using Link Communities Algorithm}
\usage{
findModules.linkcommunities(adj, nperm = 10, path, min.module.size = 30)
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function wraps a function iteratively to get modules from network adjacency
matrix using Link communities algorithm.
}
