% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.louvain.R
\name{findModules.louvain}
\alias{findModules.louvain}
\title{Find Modules with Network Adjacency Matrix Using Louvain Clustering}
\usage{
findModules.louvain(adj, nperm = 10, path, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{path}{File path location of CFinder.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function wraps a function iteratively to get modules from network adjacency
matrix using igraph's louvain clusting function.
}
