% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.megena.R
\name{findModules.megena}
\alias{findModules.megena}
\title{Find Modules with Megena Clustering}
\usage{
findModules.megena(g, n_cores = 1, alpha.cut = 1, ...)
}
\arguments{
\item{g}{An \code{igraph} graph}

\item{n_cores}{Optional. The number of cores to use for parallel computing.
If n_cores = 1, MEGENA will not compute in parallel.}

\item{alpha.cut}{Optional. Resolution cut-off for cutting the cluster
dendrogram.}

\item{...}{Optional. Additional arguments for \code{MEGENA::do.MEGENA} or
\code{MEGENA::get.union.cut}.}
}
\value{
A named vector where the names are genes and the values are cluster
  assignments
}
\description{
This function finds modules from a network adjacency matrix using MEGENA.
}
