% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.R
\name{findModules}
\alias{findModules}
\title{Find Modules TODO
TODO WGCNA and any other algorithms used in construction that also do modules}
\usage{
findModules(adj, method, nperm = 10, min.module.size = 30, n_cores = 1, ...)
}
\arguments{
\item{adj}{An n x n upper triangular adjacency matrix where "n" is the number
of genes.}

\item{method}{Which method to use to find modules. Current options are:
fast_greedy, infomap, label_prop, leading_eigen, linkcommunities, louvain,
megena, spinglass, walktrap}

\item{nperm}{Optional. Number of permutations on the gene ordering.}

\item{min.module.size}{Optional. Integer between 1 and n genes.}

\item{n_cores}{Optional. For "megena" only, how many cores to use when
computing in parallel. If n_cores = 1, megena will not compute in parallel.}

\item{...}{Optional. Additional arguments to be passed to the individual
clustering functions}
}
\value{
An n x 2 data frame with columns for "gene" and "module", where n is
  the number of genes
}
\description{
Find Modules TODO
TODO WGCNA and any other algorithms used in construction that also do modules
}
