% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.spinglass.once.R
\name{findModules.spinglass.once}
\alias{findModules.spinglass.once}
\title{Find Modules with Network Adjacency Matrix Using spinglass Clustering}
\usage{
findModules.spinglass.once(adj, min.module.size)
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function tries to get modules from network adjacency matrix using igraph's
spinglass clusting function.
}
