% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.spinglass.R
\name{findModules.spinglass}
\alias{findModules.spinglass}
\title{Find Modules with Network Adjacency Matrix Using Spinglass Clustering}
\usage{
findModules.spinglass(adj, nperm = 10, path, min.module.size = 30)
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function wraps a function iteratively to get modules from network adjacency
matrix using igraph's spinglass clusting function.
}
