% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmetanet.R
\name{ggmetanet}
\alias{ggmetanet}
\title{ggmetanet}
\usage{
ggmetanet(
  metanetwork,
  g = NULL,
  beta = 0.1,
  legend = NULL,
  mode = "TL-tsne",
  edge_thrs = NULL,
  layout_metaweb = F,
  nrep_ly = 1,
  flip_coords = F,
  diff_plot_bool = F,
  alpha_per_group = NULL,
  alpha_per_node = NULL,
  alpha_interactive = F,
  ggnet.config = ggnet.default,
  TL_tsne.config = TL_tsne.default
)
}
\arguments{
\item{metanetwork}{object of class metanetwork}

\item{g}{network (igraph object) to represent, default is metaweb}

\item{beta}{the diffusion parameter of the diffusion kernel, a positive scalar controlling the
vertical squeezing of the network}

\item{legend}{resolution for the legend, legend resolution must be a coarser resolution than the resolution of g, default is NULL}

\item{mode}{mode used for layout, 'TL-tsne' for trophic level t-sne and 'TL-kpco' for trophic level kernel based pco. Default is 'TL-tsne'}

\item{edge_thrs}{if non-null, a numeric (between 0 and 1) indicating an edge threshold for the representation}

\item{layout_metaweb}{a boolean indicating whether the layout of the metaweb should be used to represent the network
to use metaweb layout = T, you need first to compute metaweb layout for this beta value using \code{attach_layout()}}

\item{nrep_ly}{If several layouts for this beta value are attached to the metaweb (if \code{layout_metaweb = T}), index of the layout to use, see \code{attach_layout()}}

\item{flip_coords}{a boolean indicating wheter coordinates should be flipped.}

\item{diff_plot_bool}{boolean, do not edit by hand}

\item{alpha_per_group}{controlling alpha per group (only for 'ggnet' vis), a list of format
\code{list(resolutions = "XX",groups = XX,alpha_focal = XX,alpha_hidden = XX)}, see example}

\item{alpha_per_node}{controlling alpha per node (only for 'ggnet' vis), a list of format
\code{list(nodes = XX,alpha_focal = XX,alpha_hidden = XX)}, see example
In that case, y-axis is the trophic level and x-axis is the layout axis}

\item{alpha_interactive}{a boolean indicating whether alpha (that is node transparency)
should be asked in interactive mode to the user}

\item{ggnet.config}{configuration list for ggnet representation, default is ggnet.default}

\item{TL_tsne.config}{configuration list for mode 'TL-tsne', default is TL_tsne.default}
}
\value{
object of class 'ggplot'
}
\description{
Function that provides network static representation  (using 'ggnet') from a
'metanetwork' object with a layout based on a diffusion kernel
}
\examples{
library(metanetwork)
library(igraph)

g = make_ring(5,directed = TRUE)
meta0 = build_metanet(g)
meta0 = compute_TL(meta0)
ggmetanet(meta0)

# angola dataset
meta_angola = compute_TL(meta_angola)
ggmetanet(meta_angola,legend = 'Phylum',beta = 0.05)

}
