% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmetanet.R
\docType{data}
\name{ggnet.default}
\alias{ggnet.default}
\title{Default configuration for ggnet}
\format{
An object of class \code{metanetwork_config} of length 16.
}
\usage{
ggnet.default
}
\description{
A list with parameters customizing ggmetanet representation (see ggnet documentations)
}
\examples{
# display all default settings
ggnet.default

# create a new settings
ggnet.custom = ggnet.default
ggnet.custom$edge.size = 2
ggnet.custom

}
\keyword{datasets}
