% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout.R
\docType{data}
\name{group_layout.default}
\alias{group_layout.default}
\title{Default configuration for group-TL-tsne layout}
\format{
An object of class \code{list} of length 3.
}
\usage{
group_layout.default
}
\description{
A list with parameters customizing group-TL-tsne layout
}
\examples{
# display all default settings
group_layout.default

# create a new settings object with n_neighbors set to 5
group_layout.custom = group_layout.default
group_layout.custom$group_height = 10
group_layout.custom

}
\keyword{datasets}
