% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAracne.R
\name{installAracne}
\alias{installAracne}
\title{This function installs the ARACNE system Function}
\usage{
installAracne(tool_storage_loc)
}
\arguments{
\item{tool_storage_loc}{Required. Provides the directory to temporarily store
the ARACNE files and package.}
}
\value{
the file path to the installation folder for ARACNE
}
\description{
This function automatically installs ARACNE from `inst/aracne`.
}
