% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metanet.R
\name{is.metanetwork}
\alias{is.metanetwork}
\alias{is.metanetwork.metanetwork}
\title{Test of belonging to class metanetwork}
\usage{
is.metanetwork(metanetwork)

\method{is.metanetwork}{metanetwork}(metanetwork)
}
\arguments{
\item{metanetwork}{the object to test}
}
\value{
a boolean indicating whether the object belongs to class \code{metanetwork}

\code{NULL}
}
\description{
Return a boolean indicating whether the object belongs to class \code{metanetwork}
}
\examples{
library(metanetwork)
library(igraph)
 
g = make_ring(5,directed = TRUE)
meta = build_metanet(g)
is.metanetwork(meta)
#on Angola dataset
data("meta_angola")
is.metanetwork(meta_angola)
}
