% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetIC.R
\name{lassoIC}
\alias{lassoIC}
\title{Best AIC or BIC solution for Lasso regression}
\usage{
lassoIC(x, y, ...)
}
\arguments{
\item{x}{A design matrix where rows are samples and columns are genes. Can be
in sparse or dense matrix format.}

\item{y}{The response variable (a single gene), an nx1 matrix where n = the
number of rows in x (the number of samples).}

\item{...}{Optional. Other arguments accepted by \code{glmnet()}}
}
\value{
a 2 x n_genes matrix, where row 1 is the coefficients for the
  solution with the best AIC, and row 2 is for the best BIC. The rows are
  named "AIC" and "BIC".
}
\description{
This function is a wrapper for \code{glmnetIC} that runs a glmnet() function
with lasso regression and returns the results with the best AIC or BIC.
}
\seealso{
[glmnetIC]
}
