% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{loadCSVFile}
\alias{loadCSVFile}
\title{Load CSV File using data.table}
\usage{
loadCSVFile(filename, return_matrix = TRUE, ...)
}
\arguments{
\item{filename}{The path to the file to load}

\item{return_matrix}{Optional. If \code{TRUE}, the loaded data object will be
coerced to a matrix. If \code{FALSE}, the data object will be a \code{data.frame}.}

\item{...}{Optional, other arguments to pass to \code{fread()}.}
}
\value{
either a matrix or a data.frame, depending on the value of \code{return_matrix}
}
\description{
This function loads data from a CSV file using \code{data.table::fread},
which is much faster than \code{read.table} or \code{read.csv}. The data can
optionally be coerced to a matrix or left as a data.frame.
}
