% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_angola.R
\docType{data}
\name{meta_angola}
\alias{meta_angola}
\title{Angola fishery metanetwork
metanetwork built from:
Angelini & Velho 2011, Data from:
Angelini, R., Velho, VF. (2011)
Ecosystem structure and trophic analysis of Angolan fishery landings. Scientia Marina 75(2)}
\format{
\describe{
A object of class 'metanetwork'
\item{metaweb}{The metaweb from Angelini & Velho 2011, containing 28 groups and 127 interactions, a \code{igraph} object}
\item{abTable}{Abundance table built from biomass at two dates: 1986 and 2003, a \code{matrix}}
\item{trophicTable}{Taxonomic table, a three column \code{data.frame}
with three different taxonomic levels (species (or group), phylum and kingdom)}
}
}
\source{
\url{https://scientiamarina.revistas.csic.es/index.php/scientiamarina/article/view/1254}
}
\usage{
data(meta_angola)
}
\description{
Angola fishery metanetwork
metanetwork built from:
Angelini & Velho 2011, Data from:
Angelini, R., Velho, VF. (2011)
Ecosystem structure and trophic analysis of Angolan fishery landings. Scientia Marina 75(2)
}
\keyword{datasets}
