% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_norway.R
\docType{data}
\name{meta_norway}
\alias{meta_norway}
\title{Norway soil metanetwork
metanetwork built from:
Calderon-Sanou et al. 2021, Data from:
Calderon-Sanou, I., Munkemuller, T., Zinger, L., Schimann, H., Yoccoz, N. G., Gielly, L., ... & Thuiller, W. (2021).
Cascading effects of moth outbreaks on subarctic soil food webs. Scientific reports, 11(1), 1-12.}
\format{
\describe{
A object of class 'metanetwork'
\item{metaweb}{The metaweb from Calderon-Sanou et al. 2021, containing 40 groups and 204 interactions, a \code{igraph} object}
\item{abTable}{Abundance table built from eDNA data in disturbed (moth outbreaks) and non-disturbed sites, a \code{matrix}}
\item{trophicTable}{Trophic table, a three column \code{data.frame}
with three different taxonomic levels (trophic_group, trophic_class and taxa)}
}
}
\source{
\url{https://www.nature.com/articles/s41598-021-94227-z}
}
\usage{
data(meta_norway)
}
\description{
Norway soil metanetwork
metanetwork built from:
Calderon-Sanou et al. 2021, Data from:
Calderon-Sanou, I., Munkemuller, T., Zinger, L., Schimann, H., Yoccoz, N. G., Gielly, L., ... & Thuiller, W. (2021).
Cascading effects of moth outbreaks on subarctic soil food webs. Scientific reports, 11(1), 1-12.
}
\keyword{datasets}
