% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanet_build_pipe.R
\name{metanet_build_pipe}
\alias{metanet_build_pipe}
\alias{metanet_build_pipe.metanetwork}
\title{Build and execute 'metanetwork' pipeline}
\usage{
metanet_build_pipe(metanetwork, beta = 0.1, verbose = T)

\method{metanet_build_pipe}{metanetwork}(metanetwork, beta = 0.1, verbose = T)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{beta}{the diffusion parameter of the diffusion kernel, a positive scalar controlling the 
squeezing of the network}

\item{verbose}{a boolean indicating wheter message along the pipeline should be printed}
}
\value{
object of class 'metanetwork'

\code{NULL}
}
\description{
Method executing the whole metanetwork pipeline, including building 'metanetwork' object (\code{build_metanet},\code{append_agg_nets}, \code{compute_TL}, 
\code{attach_layout})
}
\examples{
library(metanetwork)
library(igraph)

g = make_lattice(dimvector = c(4,4),2,3,directed = TRUE)
meta0 = build_metanet(g)
meta0 = metanet_pipe(meta0)
ggmetanet(meta0)

#on angola data set
meta_angola = metanet_pipe(meta_angola,beta = 0.05)

}
