\name{MetaNetwork-internal}

\alias{timeToOutfile}

\title{Internal MetaNetwork functions}

\description{
  Internal MetaNetwork functions. These are generally not to be called by the user.
}

\usage{
  timeToOutfile(t1, t2, outputdir, outfile)
}


\author{Danny Arens, \email{DannyArends@Gmail.com} }

\keyword{internal}
