\name{metanetwork-package}
\alias{metanetwork-package}
\alias{metanetwork}
\docType{package}
\title{metanetwork: Statistical network exploration toolkit}
\description{Package to run multiple different statistical network inference methodologies for ultra-high dimensional data and extract topological network features.}
\details{
\tabular{ll}{
Package: \tab metanetwork\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-20\cr
License: \tab GPLv3\cr
}
TODO
}
\author{
Ben Logsdon
Maintainer: Ben Logsdon <ben.logsdon@sagebase.org>
}
\references{
references TODO
}
\keyword{ package }
