% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrnetWrapper.R
\name{mrnetWrapper}
\alias{mrnetWrapper}
\title{Wrapper for MRNET, ARACNE, and CLR from `parmigene`}
\usage{
mrnetWrapper(data, ...)
}
\arguments{
\item{data}{Required. A gene expression matrix with rows as sample IDs and
columns as Gene or feature IDs. The matrix will be transposed to work with
`parmigene`.}

\item{...}{Optional. Additional arguments accepted by either `parmigene::knnmi.all`
or `parmigene::aracne.m`.}
}
\value{
a named list of three matrices, where "aracne" is the output of
`parmigene::aracne.m`, "clr" is the output of `parmigene::clr`, and "mrnet"
is the output of `parmigene::mrnet`.
}
\description{
This function uses `parmigene` to calculate network matrices using the MRNET,
ARACNE, and CLR algorithms.
}
\details{
A mutual information matrix is calculated with `parmigene::knnmi.all`, and
then three network matrices are calculated from this mutual information
matrix using MRNET, ARACNE, and CLR. The results from all three algorithms
are returned as a named list.
}
