% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Required. Multiple ggplot objects, or a plotlist as a list of
ggplot objects}

\item{plotlist}{Optional. A list of ggplot objects. (Default = NULL)}

\item{file}{Optional. A file name to save the plot as, currently non-functional.
(Default=NULL)}

\item{cols}{Optional. (Default = 1)}

\item{layout}{Optional. The plot layout as a matrix. eg. if
\code{layout = matrix(c(1,2,3,3), nrow=2, byrow=TRUE)}, then plot 1 will go in the
upper left, 2 will go in the upper right, and 3 will go all the way across
the bottom.}
}
\value{
A ggplot image
}
\description{
This function plots multiple ggplot objects passed in as either ..., or to plotlist
(as a list of ggplot objects).
}
