% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelNetworkWrapper.R
\name{parallelNetworkWrapper}
\alias{parallelNetworkWrapper}
\title{Parallel Network Wrapper}
\usage{
parallelNetworkWrapper(
  data,
  n_cores = 0,
  cluster_type = "FORK",
  regressionFunction,
  outputpath,
  eigen = NULL,
  regulatorIndex = NULL,
  debug_save = FALSE
)
}
\arguments{
\item{data}{Required. Expression matrix to be used for network construction,
which should have samples as rows and genes as columns.}

\item{n_cores}{Optional. The number of parallel cores to use.}

\item{cluster_type}{Optional. Use "FORK" if running on a Unix system, and
"PSOCK" if running on Windows.}

\item{regressionFunction}{Required. The name of the regression function to use,
which should exactly match the name of a function in this package. Current
options are: lassoAIC, lassoBIC, lassoCV1se, lassoCVmin, ridgeAIC, ridgeBIC,
ridgeCV1se, ridgeCVmin, sparrowZ, sparrow2Z}

\item{outputpath}{Required. The output path to save the resulting coexpression
network}

\item{eigen}{Required for ridgeAIC and ridgeBIC, ignored otherwise. Eigen
cutoff value.}

\item{regulatorIndex}{Optional. A vector of numerical indexes into
`colnames(data)` for a subset of genes that should be used in the network,
rather than using the full gene set.}

\item{debug_save}{Optional. Whether to save the network matrix as an RDS
file, for debugging purposes.}
}
\value{
a vector of file names corresponding to the files generated by this
function. Currently the only file is: `<outputpath>/<regressionFunction>Network.csv`
}
\description{
Wrapper to run network detection in a parallel fashion
}
