% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\description{
Like dplyr, metanetwork also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{

library(metanetwork)
meta_angola \%>\% attach_layout() \%>\% ggmetanet()
}
