% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trophicTable.R
\name{plot_trophicTable}
\alias{plot_trophicTable}
\title{Plot trophic groups hierarchy}
\usage{
plot_trophicTable(metanetwork, res = "all", ggnet.config = ggnet.default)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{res}{resolutions included in the hierarchy representation. Default is "all" (all resolutions are then included)
but can be also a vector of given resolutions}

\item{ggnet.config}{configuration list for ggnet representation, default is ggnet.default}
}
\value{
object of class 'ggnet', representation of group hierarchy
}
\description{
Function to represent trophic groups hierarchy provided by trophicTable
}
\examples{
library(metanetwork)

#on Angola data_set
data("meta_angola")
plot_trophicTable(meta_angola)

}
