% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.metanetwork}
\title{print metanetwork}
\usage{
print(metanetwork)

\method{print}{metanetwork}(metanetwork)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}
}
\value{
character indicating number of nodes and edges of the metaweb, available resolutions
and number of local networks

\code{NULL}
}
\description{
Print method for class \code{metanetwork}
}
\examples{
library(metanetwork)
library(igraph)

g = make_ring(5,directed = TRUE)
meta = build_metanet(g)
print(meta)

#on Angola dataset
data("meta_angola")
print(meta_angola)

#on Norway dataset
data("meta_norway")
print(meta_norway)
}
