% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_metanet.R
\name{print_metanet}
\alias{print_metanet}
\alias{print_metanet.metanetwork}
\title{print metanetwork}
\usage{
print_metanet(metanetwork)

\method{print_metanet}{metanetwork}(metanetwork)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}
}
\value{
character indicating number of nodes and edges of the metaweb, availabe resolutions 
and number of local networks

\code{NULL}
}
\description{
Print method for class \code{metanetwork}
}
\examples{
library(metanetwork)
library(igraph)

g = make_ring(5,directed = TRUE)
meta = build_metanet(g)
print_metanet(meta)

#on Angola dataset
print_metanet(meta_angola)

}
