% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneBICnetwork.R
\name{pruneBICnetwork}
\alias{pruneBICnetwork}
\title{Prune a network using a BIC Curve}
\usage{
pruneBICnetwork(network, exprData, maxEdges = 2e+05)
}
\arguments{
\item{network}{A network object}

\item{exprData}{Expression matrix where rows are samples and columns are genes}

\item{maxEdges}{Optional. Maximum number of edges to use.}
}
\value{
A named list containing: "network" = a matrix containing the
  original network subset to the edges that produce the smallest BIC,
  "bicMin" = the smallest BIC value, "bicPath" = a vector containing the BIC
  values at each iteration.
}
\description{
This function gets a list of edges from the input network, creates a new
empty network, iteratively adds edges to the new network, and computes the BIC
after adding each edge. The final network will be the set of edges that
produce the smallest BIC.
}
