% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankConsensus2.R
\name{rankConsensus2}
\alias{rankConsensus2}
\title{Ranks Consensus networks with ties.}
\usage{
rankConsensus2(networks)
}
\arguments{
\item{networks}{Required. A list object containing an individual network as 
a list entry.}
}
\value{
The best rank consensus network.
}
\description{
This function ranks a list of consensus network objects and returns the best 
rank network. This version of rank consensus is compatible with ties.
}
