% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatorAnalysis.directed.R
\name{regulatorAnalysis.directed}
\alias{regulatorAnalysis.directed}
\title{Function to Identify Network Regulators from Directed Networks}
\usage{
regulatorAnalysis.directed(adj, G, h = 3, FDR = 0.05)
}
\arguments{
\item{adj}{Required. An n x n weighted upper triangular adjacency in the matrix 
class format.}

\item{G}{Required. A named vector of node scores.}

\item{h}{Optional. Neighborhood search distance (h nodes away from current node) 
(Default = 3)}

\item{FDR}{Optional. Adjusted pvalue cutoff for regulator selection.
(Default = 0.05)}
}
\value{
scores = n x 5 dimensional data frame with columns giving neighborhood
based score, adjusted pvalue, whether a gene is regulator/global regulator.
}
\description{
Identifies network un-weighted regulators from directed networks
}
