% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatorAnalysis.directed_weighted.R
\name{regulatorAnalysis.directed_weighted}
\alias{regulatorAnalysis.directed_weighted}
\title{Function to Identify Network Regulators from Directed Weighted Networks}
\usage{
regulatorAnalysis.directed_weighted(
  adj,
  G,
  h = 3,
  n = 100,
  correction.method = "bonferroni",
  pval.cutoff = 0.01
)
}
\arguments{
\item{adj}{Required. An n x n weighted upper triangular adjacency in the matrix 
class format.}

\item{G}{Required. A named vector of node scores.}

\item{h}{Optional. Neighborhood search distance (h nodes away from current node) 
(Default = 3)}

\item{n}{Optional. Number of randomisation for pvalue computation. (Default = 100)}

\item{correction.method}{Optional. Multiple testing correction method. Options
are; c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")
(Default = 'bonferroni')}

\item{pval.cutoff}{Optional. Adjusted pvalue cutoff for regulator selection.
(Default = 0.01)}
}
\value{
scores = n x 5 dimensional data frame with columns giving neighborhood
based score, adjusted pvalue, whether a gene is regulator/global regulator.
}
\description{
Identifies network regulators from directed network weights.
}
