% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatorAnalysis.undirected_weighted.R
\name{regulatorAnalysis.undirected_weighted}
\alias{regulatorAnalysis.undirected_weighted}
\title{Function to Identify Network Regulators from Undirected Weighted Networks}
\usage{
regulatorAnalysis.undirected_weighted(adj, G, h = 3, n = 100, FDR = 0.05)
}
\arguments{
\item{adj}{Required. An n x n weighted upper triangular adjacency in the matrix 
class format.}

\item{G}{Required. A named vector of node scores.}

\item{h}{Optional. Neighborhood search distance (h nodes away from current node) 
(Default = 3)}

\item{n}{Optional. number of permuations to run. (Default = 100)}

\item{FDR}{Optional. Adjusted pvalue cutoff for regulator selection.
(Default = 0.05)}
}
\value{
scores = n x 4 dimensional list with columns giving neighborhood
based score, adjusted pvalue, whether a gene is regulator/global regulator.
}
\description{
Identifies network un-weighted regulators from Weighted undirected networks
}
