% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.consensus.cluster.R
\name{run.consensus.cluster}
\alias{run.consensus.cluster}
\title{Runs Consensus Clustering Algorithm}
\usage{
run.consensus.cluster(
  d,
  kGrid = NULL,
  repCount = NULL,
  diss = inherits(d, "dist"),
  pItem = NULL,
  pFeature = NULL,
  innerLinkage = NULL,
  distance = NULL,
  clusterAlg = NULL,
  weightsItem = NULL,
  weightsFeature = NULL,
  verbose = NULL,
  corUse = NULL
)
}
\arguments{
\item{d}{Required. A matrix where columns=items/samples and rows are features.
For example, a gene expression matrix of genes in rows and microarrays in columns.
OR ExpressionSet object. (Default = NULL)}

\item{kGrid}{Optional. (Default = NULL)}

\item{repCount}{Optional. Replicate count. (Default = NULL)}

\item{diss}{Optional. A distance matrix of d. (Default = inherits( d, "dist" ))}

\item{pItem}{Optional. A numerical value. proportion of items to sample.
(Default = 0.8)}

\item{pFeature}{Optional. A numerical value. proportion of features to sample.
(Default = 1)}

\item{innerLinkage}{Optional. A heirarchical linkage method for subsampling.
(Default = "average")}

\item{distance}{Optional. A character value. sample distance measures:
"pearson","spearman", or "euclidean". (Default = "pearson")}

\item{clusterAlg}{Optional. A character value. cluster algorithm. "hc"
heirarchical (hclust) or "km" for kmeans. (Default = "kmeans")}

\item{weightsItem}{Optional. A numerical vector. weights to be used for
sampling items. (Default = NULL)}

\item{weightsFeature}{Optional. AN umerical vector. weights to be used for
sampling features. (Default = NULL)}

\item{verbose}{Optional. A boolean when set to TRUE, prints messages to the
screen to indicate progress. This is useful for large datasets.(Default = FALSE)}

\item{corUse}{Optional. Use all cores avaiable. (Default = "Everything")}
}
\value{
Clustered consensus matrix. areaUnderCDF = areaK,
consensus.matrix = cns.mtrx
}
\description{
A modiefed parallel version of code imported from
https://github.com/Bioconductor-mirror/ConsensusClusterPlus 1.11.1. Taken out
of findModules.consensusCluster.R
}
