% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_metanetwork_app}
\alias{run_metanetwork_app}
\title{Function to run the shiny app to generate the figure}
\usage{
run_metanetwork_app()
}
\description{
Function to run the shiny app to generate the figure
}
