% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.nodes.R
\name{score.nodes}
\alias{score.nodes}
\title{Scores Nodes From Regulator Discovery}
\usage{
score.nodes(g, G, h = 3, mode = "all")
}
\arguments{
\item{g}{Required. An igraph object with n vertices}

\item{G}{Required. A named vector of node scores.}

\item{h}{Optional. Neighborhood search distance (h nodes away from current node)
(Default = 3)}

\item{mode}{Optional. One of c("all", "out", "in", "total"). Character string, 
"out" for out-degree, "in" for in-degree or "all" for the sum of the two. 
For undirected graphs this argument is ignored. (Default = 'all')}
}
\value{
node.scores = n x 1 dimensional vector of node scores based on its
neighborhood
}
\description{
Function to scores results of regulator scoring analysis.
}
