% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparrow2Z.R
\name{sparrow2Z}
\alias{sparrow2Z}
\title{Runs vbsr with a 2Z cutoff for a gene across a matrix}
\usage{
sparrow2Z(y, x, fdr = NULL, ...)
}
\arguments{
\item{y}{Required. response variable. Normally distributed errors for
family="normal". For family="binomial" should be coded as a vector of 0's and 1's.}

\item{x}{Required. Design matrix, an n x m matrix, with rows as observations.}

\item{fdr}{Optional. FDR threshold cut off for edge determination. NULL results
in a cutoff of 0.05 (Default = NULL)}
}
\value{
A coexpression value
}
\description{
This function wraps variable bays spike regression of a genes expression across
a matrix of genes expressed in the same samples with a 2Z cutoff value.
}
