% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparrowNetwork.R
\name{sparrowNetwork}
\alias{sparrowNetwork}
\title{Runs wraps vbsr across a matrix}
\usage{
sparrowNetwork(data)
}
\arguments{
\item{data}{Required. A .}
}
\value{
A coexpression matrix
}
\description{
This function wraps variable bays spike regression calling all vsbr on all genes 
pairwise across an expression matrix
}
