% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.igraphWrapper.R
\name{spinglass_wrapper}
\alias{spinglass_wrapper}
\title{Spinglass wrapper}
\usage{
spinglass_wrapper(g, min.module.size = 30, ...)
}
\arguments{
\item{g}{An \code{igraph} graph}

\item{min.module.size}{Optional. The minimum number of genes per module,
which is used here to avoid running spinglass on connected graphs with
fewer than \code{min.module.size} genes in them.}

\item{...}{Optional. Other arguments accepted by \link[igraph:cluster_spinglass]{igraph::cluster_spinglass}.}
}
\value{
a named vector where the names are gene names and the values are
module membership.
}
\description{
This function wraps \link[igraph:cluster_spinglass]{igraph::cluster_spinglass} so it can be run on all
connected sub-graphs of the main graph.
}
\details{
\link[igraph:cluster_spinglass]{igraph::cluster_spinglass} only works on connected graphs, and the main
graph may have components that are not connected to other components. This
function divides the main graph into sub-graphs that are connected, runs
\code{cluster_spinglass} on each sub-graph, and concatenates the membership
results into one vector.
}
