% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synGetFiles.R
\name{synGetFiles}
\alias{synGetFiles}
\title{This function pulls files from a synapse project}
\usage{
synGetFiles(project_id, pattern_id, downloadLocation = getwd())
}
\arguments{
\item{project_id}{Required. A character vector of a synapse ID of a synapse project.}

\item{pattern_id}{Required. A character vector too match in the file names of 
the children entities in `project_id`.}

\item{downloadLocation}{Optional. Local directory to download the files to. 
Default = `getwd()`}
}
\value{
A character vector of file paths
}
\description{
This function downloads all of the network files in a synapse parent folder using
the synID of the folder containing folders of specific network types. It uses
the name of the folder plus a pattern id eg. 'Network' to find network files. 
for example the `ridgeAIC/` folder in `project_id` will need to contain the
`ridgeAICNetwork` file if pattern_id = 'Network' and `project_id` is set to 
the parent folder containing `ridgeAIC/`. The files will be downloaded to the
path specified in `downloadLocation`.
}
