% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismetaNetwork.R
\docType{data}
\name{visNetwork.default}
\alias{visNetwork.default}
\title{Default configuration for visNetwork}
\format{
An object of class \code{metanetwork_config} of length 4.
}
\usage{
visNetwork.default
}
\description{
A list with parameters customizing visNetwork visualisation (see visNetwork documentations)
}
\examples{
# display all default settings
visNetwork.default

# create a new settings
visNetwork.custom = visNetwork.default
visNetwork.custom$label.size = 10
visNetwork.custom

}
\keyword{datasets}
