% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcnaTOM.R
\name{wgcnaTOM}
\alias{wgcnaTOM}
\title{Runs WGCNA TOMsimilarity}
\usage{
wgcnaTOM(
  data,
  outputpath,
  RsquaredCut = 0.85,
  defaultPower = 6,
  ncores = 0,
  debug_save = FALSE
)
}
\arguments{
\item{data}{Required. A Matrix containing gene expression values with samples
as row values and genes as column IDs}

\item{outputpath}{Required. Path to save the resulting network}

\item{RsquaredCut}{Optional. Minimum R^2 value to use when picking a soft threshold.
Corresponds to the RsquaredCut argument in WGCNA::pickSoftThreshold.}

\item{defaultPower}{Optional. The default soft power to use if pickSoftThreshold
returns NA.}

\item{ncores}{Optional. How many cores/threads to allow when running WGCNA.
The default value of 0 means "do not thread".}

\item{debug_save}{Optional. Whether to save the TOM similarity matrix as an RDS
file before reducing it to an upper triangular matrix, for debugging purposes.}
}
\value{
a vector of file names corresponding to the files generated by this
function. This vector will have 3 file names: `<outputpath>/wgcnaPowerEstimate.txt`,
`<outputpath>/wgcnaSoftThresholdNetwork.csv`, and
`<outputpath>/wgcnaTopologicalOverlapMatrixNetwork.csv`
}
\description{
Runs WGCNA::TOMsimilarity over a gene expression matrix.
}
