% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcnaWrapper.R
\name{wgcnaWrapper}
\alias{wgcnaWrapper}
\title{Wrapper for WGCNA}
\usage{
wgcnaWrapper(data, n_cores = 1, ...)
}
\arguments{
\item{data}{Required. A Matrix containing gene expression values with samples
as row values and genes as column IDs}

\item{n_cores}{Optional. How many cores/threads to allow when running WGCNA.
The default value of 1 means "do not thread".}

\item{...}{Optional. Additional arguments accepted by either `WGCNA::pickSoftThreshold`,
`WGCNA::adjacency`, or `WGCNA::TOMsimilarity`.}
}
\value{
a named list of two network matrices, where "adjacency" is the output
of `WGCNA::adjacency` and "TOM" is the output of `WGCNA::TOMsimilarity`.
}
\description{
Runs WGCNA::TOMsimilarity over a gene expression matrix.
}
