% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorizeData.R
\name{winsorizeData}
\alias{winsorizeData}
\title{Winsorize a Gene Expression Matrix}
\usage{
winsorizeData(x)
}
\arguments{
\item{x}{Gene expression matrix where rows are genes and columns are samples}
}
\value{
A winsorized expression matrix, which has been transposed so that
rows are samples and columns are genes
}
\description{
Winsorizes a given gene expression matrix.
}
\details{
The top and bottom 1% of values for each gene will be replaced with the .99
and .01 quantile values for that gene, respectively. All NA values for the
gene are replaced with the mean expression for that gene. The data is then
scaled and centered on 0.
}
