% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{writeCSVFile}
\alias{writeCSVFile}
\title{Write CSV File}
\usage{
writeCSVFile(object, filename, ...)
}
\arguments{
\item{object}{The object to write to disk, which should be of a type supported
by \code{fread}. This function assumes that the columns and rows are named.}

\item{filename}{The path and name of the file where the object should be saved}

\item{...}{Optional. Additional arguments to \code{fread}.}
}
\value{
Nothing
}
\description{
Convenience wrapper for \code{data.table::fwrite}, which is faster than
\code{write.csv}.
}
