% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{writeUpperTri}
\alias{writeUpperTri}
\title{Write upper triangular matrix}
\usage{
writeUpperTri(object, filename, ...)
}
\arguments{
\item{object}{A square matrix or other object coercible to a matrix.}

\item{filename}{The path and name of the file where the object should be
saved}

\item{...}{Optional. Additional arguments to \link[data.table:fread]{data.table::fread}.}
}
\value{
Nothing
}
\description{
Convenience wrapper for \link{writeCSVFile} which converts a matrix to an upper
triangular matrix before writing to a CSV file.
}
