% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{aracne}
\alias{aracne}
\title{This function applies ARACNE on the data}
\usage{
aracne(
  data,
  path = NULL,
  pval = NULL,
  outputpath,
  na_fill = NULL,
  tool_storage_loc = NULL
)
}
\arguments{
\item{data}{Required. The gene expression matrix with rows as sample IDs and
columns as Gene or feature IDs}

\item{path}{Optional. String containing the path to the aracne compiled 
executable. (Default = NULL)}

\item{pval}{Optional. Cutoff p-value to determine a coexpressed edge. (Default = 0.05)}

\item{outputpath}{Required. The path the resulting network should be saved to.}

\item{na_fill}{Optional. Value to replace `NA` values with ideally, a large 
negative number or use min(data). (Default = NULL)}

\item{tool_storage_loc}{Oprional. Provides the directory inside docker to 
temporarily store the ARACNE files and package. (Default = config$input_profile$temp_storage_loc)}
}
\value{
A Co-Expression Network saved to the path `outputpath` and titled
 `aracneThresholdNetwork.csv` (if `pval` < 1) or as `aracneNetwork.csv` if 
 `pval` is set to 1.
}
\description{
This function takes in a gene expression matrix of gene expression and applies
the ARACNE gene co-expression network analysis frame work to find coexpressed
gene pairs in the matrix. The ARACNE framework is also installed from within
the package located in the `inst/tools/` direcctory. For more information on the 
ARACNE framework see: <https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-7-S1-S7>
}
