% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrarySparsity.R
\name{arbitrarySparsity}
\alias{arbitrarySparsity}
\title{This function Builds an arbitary sparse network}
\usage{
arbitrarySparsity(nedges, network)
}
\arguments{
\item{nedges}{Required. Edges parameter to increase sparsity.}

\item{network}{Required. A matrix or data frame of gene expression}
}
\value{
a logical matrix of TRUE/FALSE where True indicates sparsity
}
\description{
This introduces sparsity into a data frame or matrix with increasing frequency
as the `nedges` parameter is specified
}
