% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConsensus.R
\name{buildConsensus}
\alias{buildConsensus}
\title{This function builds the Consensus Network from the component network}
\usage{
buildConsensus(outputpath, networkFolderId, fileName, bar)
}
\arguments{
\item{outputpath}{Required. Local directory to load files from and save files to.}

\item{networkFolderId}{Required. The Synapse parent ID of the folder containing
the folders with the individual networks.}

\item{fileName}{Required. The input file path.
the children entities in `project_id`.}

\item{bar}{Required. File paths object with network paths}
}
\value{
Saves a rank consensus network to `outputpath` and saves the BICNetwork
 object to `outputpath` if `fileName` is specified
}
\description{
This function builds a consensus co-expression network. TODO
}
