% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConsensus.R
\name{buildConsensus}
\alias{buildConsensus}
\title{This function builds the Consensus Network from the component network}
\usage{
buildConsensus(network_files, exprData, outputpath, max_edges = 2e+05)
}
\arguments{
\item{network_files}{A vector of file paths pointing to each network matrix.}

\item{exprData}{A matrix of the gene expression data used to generate the
networks. Rows must be samples and columns must be genes.}

\item{outputpath}{Path to the folder where the consensus network should be
saved.}

\item{max_edges}{Optional. The maximum number of network edges to use when
computing the BIC curve.}
}
\value{
Nothing. Saves the rank consensus network to
  \code{<outputpath>/rankConsensusNetwork.csv} and saves the BICNetwork
  object to \code{<outputpath>/bicNetworks.rds}
}
\description{
This function builds a consensus co-expression network.
}
\details{
TODO add consensus by weight instead of rank?
TODO better save format for bicNetworks
}
