% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModularityDensity.R
\name{community_edges}
\alias{community_edges}
\title{Community Edges}
\usage{
community_edges(g, mod)
}
\arguments{
\item{g}{An \code{igraph} graph}

\item{mod}{A vector where names are genes and values are module membership}
}
\value{
A named list with names "edge.comm" and "renamed_mod". "edge.comm" =
an NxN matrix, where N is the number of modules/communities. The diagonal
will contain the number of internal edges in each module, and the
non-diagonals will contain the number of edges between module i and module
j. "renamed_mod" = \code{mod}, if \code{mod} is already a character vector, otherwise
all values of \code{mod} have an "m" pasted in front to force it to be a
character vector. This avoids ambiguity between module numbers and indexing
into arrays.
}
\description{
Get the number of edges within and between communities.
}
