% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModularityDensity.R
\name{compute.ModularityDensity}
\alias{compute.ModularityDensity}
\title{Find Global Modularity (Qds)}
\usage{
compute.ModularityDensity(adj, mod)
}
\arguments{
\item{adj}{Required. An n x n upper triangular adjacency in the matrix class
format.}

\item{mod}{Required. An n x 3 dimensional data frame with column names as
Gene.ID, moduleNumber, and moduleLabel.}
}
\value{
Qds = module density.
}
\description{
This function calculates modularity density. TODO needs review.
TODO this might be combinable with compute.Modularity to avoid calling
some of the igraph functions twice.
}
