% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModularityDensity.R
\name{compute.ModularityDensity}
\alias{compute.ModularityDensity}
\title{Find Modularity Density (Qds)}
\usage{
compute.ModularityDensity(g, mod)
}
\arguments{
\item{g}{An \code{igraph} graph of the network}

\item{mod}{A named vector where names are genes and values are module membership}
}
\value{
Qds = modularity density.
}
\description{
This function calculates modularity density according to Chen, Nguyen, and
Szymanski 2013.
}
\references{
Mingming Chen, Tommy Nguyen, Boleslaw K. Szymanski, 2013. A New
Metric for Quality of Network Community Structure.
https://doi.org/10.48550/arXiv.1507.04308
}
