% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModuleQualityMetric.R
\name{compute.ModuleQualityMetric}
\alias{compute.ModuleQualityMetric}
\title{Find  Modularity Quality}
\usage{
compute.ModuleQualityMetric(adj, mod)
}
\value{
metrics = data frame of module quality metrics.
}
\description{
This function finds qulaity metrics of all modules
}
