% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModuleQualityMetric.R
\name{compute.ModuleQualityMetric}
\alias{compute.ModuleQualityMetric}
\title{Find Modularity Quality}
\usage{
compute.ModuleQualityMetric(g, mod)
}
\arguments{
\item{g}{An \code{igraph} graph of the network}

\item{mod}{A named vector where names are genes and values are module membership}
}
\value{
metrics = data frame of module quality metrics.
}
\description{
This function finds quality metrics of all modules
}
