% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ModuleQualityMetric.R
\name{compute.ModuleQualityMetric}
\alias{compute.ModuleQualityMetric}
\title{Find  Modularity Quality}
\usage{
\method{compute}{ModuleQualityMetric}(adj, mod)
}
\arguments{
\item{adj}{Required. An n x n upper triangular adjacency in the matrix class 
format.}

\item{mod}{Required. An n x 3 dimensional data frame with column names as 
Gene.ID, moduleNumber, and moduleLabel.}
}
\value{
metrics = data frame of module quality metrics.
}
\description{
This function finds qulaity metrics of all modules
}
