% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{compute network metrics}
\usage{
compute_metrics(metanetwork, res = NULL)
}
\arguments{
\item{metanetwork}{object of class 'metanetwork'}

\item{res}{a vector containing the resolutions at which the metrics are computed}
}
\value{
a \code{data.frame}
}
\description{
Function to compute (some) network metrics on the metaweb and local networks
}
\details{
XXXXXXXXXXXXXXXXXXXXXx
}
\examples{
library(metanetwork)
library(igraph)

#on angola dataset
data("meta_angola")
compute_metrics(meta_angola)

#computing metrics only at Phylum level
compute_metrics(meta_angola,res = "Phylum")

}
\references{
Ohlmann, M., Miele, V., Dray, S., Chalmandrier, L., O'connor, L., & Thuiller, W. (2019). Diversity indices for ecological networks: a unifying framework using Hill numbers. Ecology letters, 22(4), 737-747.
}
