% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBICcurve.R
\name{computeBICcurve}
\alias{computeBICcurve}
\title{This Function Computes a Network BIC Curve}
\usage{
computeBICcurve(network, exprData, maxEdges = 2e+05)
}
\arguments{
\item{network}{A network object}

\item{exprData}{Expression matrix where rows are samples and columns are genes}

\item{maxEdges}{Optional. Maximum number of edges to use.}
}
\value{
A list object containing values of a sparse network, the best fit
minimum BIC, and TODO
}
\description{
(?) TODO
}
