% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDriverDistancePvalue.R
\name{computeDriverDistancePvalue}
\alias{computeDriverDistancePvalue}
\title{Compute Driver Distance P-Value}
\usage{
computeDriverDistancePvalue(geneSet, graph, nsamp = 100)
}
\arguments{
\item{geneSet}{Required. A user specified gene set corresponding to vertex names
in `graph`. All the names are not required to be entirely represented in the 
graph object.}

\item{graph}{Required. An igraph graph object consisting of vertices of genes
and edges representing co-expression between genes.}

\item{nsamp}{Optional. Number of permutations to build null distribution for
p-value construction (Default = 100)}
}
\value{
a list consisting of a mean distance of genes within the largest component 
to each other, the pairwise list of distances, and a vector the length of `nsamp`
consisting of the null distribution to calculate the p-value.
}
\description{
Calculates the Driver Distance of `geneSet` in `graph` as described in
`metanetworks::computeDriverDistance` and permutes `nsamp` permutations of random
gene sets from `graph` which are the size of `geneSet` to create a null distribution
to derive p-values for the Driver Distances of `geneSet`.
}
