% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.parmigene.R
\name{constructNetwork.parmigene}
\alias{constructNetwork.parmigene}
\title{Wrapper for MRNET, ARACNE, and CLR from \code{parmigene}}
\usage{
constructNetwork.parmigene(data, ...)
}
\arguments{
\item{data}{A gene expression matrix with rows as sample IDs and columns as
Gene or feature IDs. The matrix will be transposed to work with
\code{parmigene}.}

\item{...}{Optional. Additional arguments accepted by either
\link[parmigene:knnmi.all]{parmigene::knnmi.all} or \link[parmigene:aracne.m]{parmigene::aracne.m}. The \code{clr} and \code{mrnet}
functions do not accept additional arguments.}
}
\value{
a named list of three matrices, where "aracne" is the output of
\link[parmigene:aracne.m]{parmigene::aracne.m}, "clr" is the output of \link[parmigene:clr]{parmigene::clr}, and "mrnet"
is the output of \link[parmigene:mrnet]{parmigene::mrnet}.
}
\description{
This function uses \code{parmigene} to calculate network matrices using the MRNET,
ARACNE, and CLR algorithms.
}
\details{
A mutual information matrix is calculated with \link[parmigene:knnmi.all]{parmigene::knnmi.all}, and
then three network matrices are calculated from this mutual information
matrix using MRNET, ARACNE, and CLR. The results from all three algorithms
are returned as a named list.
}
