% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceSelectionMBPath.R
\name{covarianceSelectionMBPath}
\alias{covarianceSelectionMBPath}
\title{Covariance based on Neighborhood Selection}
\usage{
covarianceSelectionMBPath(X, rankedEdges, startI = 1)
}
\arguments{
\item{X}{An expression matrix where rows are samples and columns are genes.}

\item{rankedEdges}{An edge list in the form of a matrix where column one is
gene one and column 2 is gene two}

\item{startI}{Optional. Start at the first edge in `rankedEdges` (Default = 1)}
}
\value{
A list object of containing the BIC estimate, bicNeighborhood (?) , neighborhoods (?), flag (?)
}
\description{
This function defines the covariance neighborhood between gene-gene edges
in an expression matrix. TODO
}
