% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metanet.R
\name{extract_networks}
\alias{extract_networks}
\title{extract networks from a metanetwork object}
\usage{
extract_networks(metanetwork)
}
\arguments{
\item{metanetwork}{the object whose networks need to be extracted}
}
\value{
a list of \code{igraph} objects with attributes computed by \code{metanetwork}
}
\description{
Function to extract metawebs and local networks from a metanetwork object
}
\details{
Return a list of 'igraph' objects
}
\examples{
library(metanetwork)
data("meta_angola")
nets = extract_networks(meta_angola)  
sapply(nets,class)
}
