% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{fastlm}
\alias{fastlm}
\title{Fast Linear Modeling}
\usage{
fastlm(y, x)
}
\arguments{
\item{y}{Required. A vector of response values}

\item{x}{Required. A vector or matric of the same number of observations/rows
as y}
}
\value{
A vector of ZValues
}
\description{
This function retuns results from a fast linear model
}
