% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{fastlm_z}
\alias{fastlm_z}
\title{Z value for fastlm}
\usage{
fastlm_z(y, X)
}
\arguments{
\item{y}{A vector of response values}

\item{X}{A matrix of the same number of observations/rows as y}
}
\value{
A vector of Z-values
}
\description{
This function runs \code{fastlm} and returns the Z-values associated with the
solution.
}
