% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsrWrapper.R
\name{fdrThres}
\alias{fdrThres}
\title{FDR Threshold}
\usage{
fdrThres(pval, fdr = 0.05)
}
\arguments{
\item{pval}{A vector of uncorrected p-values.}

\item{fdr}{Optional. Desired FDR cutoff. (Default = 0.05)}
}
\value{
A corrected threshold for p-value cutoff
}
\description{
This function calculates a corrected FDR threshold for p-values, which is
corrected based on the number of p-values.
}
