% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.CFinder.once.R
\name{findModules.CFinder.once}
\alias{findModules.CFinder.once}
\title{This function tries to get modules from network adjacency matrix using Link 
communities algorithm.}
\usage{
findModules.CFinder.once(adj, path, min.module.size, i)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}

\item{i}{Required. Character name of temp file name to generate.}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function tries to get modules from network adjacency matrix using Link 
communities algorithm.
}
