% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.consensusCluster.R
\name{findModules.consensusCluster}
\alias{findModules.consensusCluster}
\title{Finds Consensus Clusters}
\usage{
findModules.consensusCluster(
  d = NULL,
  maxK = 100,
  reps = 100,
  pGenes = 0.8,
  clusterAlg = "kmeans",
  hclust_method = "average",
  distance = "pearson",
  changeCDFArea = 0.001,
  nbreaks = 20,
  seed = 101,
  corUse = "everything",
  verbose = F
)
}
\arguments{
\item{d}{Optional. A matrix where columns=items/samples and rows are features.
For example, a gene expression matrix of genes in rows and microarrays in columns.
OR ExpressionSet object. (Default = NULL)}

\item{maxK}{Optional. An  integer value. maximum cluster number to evaluate.
(Default = 100)}

\item{reps}{Optional. An integer value. number of subsamples.  (Default = 100)}

\item{pGenes}{Optional. A numerical value. proportion of items to sample.
(Default = 0.8)}

\item{clusterAlg}{Optional. A character value. cluster algorithm. "hc"
heirarchical (hclust) or "km" for kmeans. (Default = "kmeans")}

\item{hclust_method}{Optional. The "method" argument of \code{hclust}}

\item{distance}{Optional. A character value. sample distance measures:
"pearson", "spearman", or "euclidean". (Default = "pearson")}

\item{changeCDFArea}{Optional. Minimum spline distance for seq(2,`maxK`,
length.out = `nbreaks`) (Default = 0.001)}

\item{nbreaks}{Optional. Number of breaks to use in
seq(2,`maxK`,length.out = `nbreaks`) this becomes the kGrid argument in
run.consensus.cluster. (Default = 20)}

\item{seed}{Optional, A numerical value. Sets random seed for reproducible results.}

\item{corUse}{Optional. Use all cores avaiable. (Default = "Everything") TODO this is wrong}

\item{verbose}{Optional. A boolean when set to TRUE, prints messages to the
screen to indicate progress. This is useful for large datasets.(Default = FALSE)}
}
\value{
Final clustered modules.

TODO does this really need to be run on partition.adj? Wouldn't it make sense
to do the consensus matrix like in run.consensus.cluster using all the different
module results instead?
}
\description{
A modified parallel version of code imported from
https://bioconductor.org/packages/ConsensusClusterPlus 1.70.0
}
