% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.consensusCluster.R
\name{findModules.consensusCluster}
\alias{findModules.consensusCluster}
\title{Finds Consensus Clusters}
\usage{
findModules.consensusCluster(
  d,
  maxK = 100,
  n_subsamples = 100,
  pGenes = 0.8,
  clusterAlg = "hclust",
  hclust_method = "average",
  distance_metric = "pearson",
  changeCDFArea = 0.001,
  n_ks = 20,
  corUse = "everything",
  n_cores = 1,
  log_file_path = ".",
  verbose = FALSE,
  seed = 101
)
}
\arguments{
\item{d}{An MxN matrix, where M is the number of clustering methods * the
number of clusters in each method, N is the number of genes, and the values
represent cluster membership: 1 if a gene is in that cluster, 0 if not.}

\item{maxK}{Optional. Maximum number of clusters to evaluate}

\item{n_subsamples}{Optional. Number of random sub-samples of the genes to
use.}

\item{pGenes}{Optional. Proportion of genes to sample at each random
sub-sample.}

\item{clusterAlg}{Optional. One of three cluster algorithms: "hclust",
"kmeans", or "pam".}

\item{hclust_method}{Optional. The "method" argument of \link{hclust}, if
\code{clusterAlg} is "hclust"}

\item{distance_metric}{Optional. The metric to use to calculate distances:
"pearson", "spearman", or "euclidean".}

\item{changeCDFArea}{Optional. Minimum spline distance for seq(2,\code{maxK},
length.out = \code{n_ks})}

\item{n_ks}{Optional. Rather than test every \code{k} in the range (2, \code{maxK}),
only \code{n_ks} values in the range will be tested, chosen with \code{seq(2, maxK, length.out = n_ks)}.}

\item{corUse}{Optional. The "use" argument of the \link[stats:cor]{stats::cor} function. Only
used when \code{clusterAlg} is "hclust" or "pam" and \code{distance_metric} is
"pearson" or "spearman".}

\item{n_cores}{Optional. Number of parallel cores to use when computing
clusters. If \code{n_cores} = 1, the function will not run in parallel.}

\item{log_file_path}{Optional. If running in parallel, output will be written
to a log file in the directory specified by \code{log_file_path}. By default,
logs are written to the current working directory.}

\item{verbose}{Optional. When set to TRUE, prints messages to the screen to
indicate progress. When FALSE, no messages are printed.}

\item{seed}{Optional. The number to use to set the random seed, for
reproducible results.}
}
\value{
Final clustered modules. TODO

TODO does this really need to be run on partition.adj? Wouldn't it make sense
to do the consensus matrix like in run.consensus.cluster using all the different
module results instead?
}
\description{
A modified parallel version of code imported from
https://bioconductor.org/packages/ConsensusClusterPlus 1.70.0
}
