% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.consensusCluster.R
\name{findModules.consensusCluster}
\alias{findModules.consensusCluster}
\title{Finds Consensus Clusters}
\usage{
findModules.consensusCluster(
  d = NULL,
  maxK = 100,
  reps = 100,
  pItem = 0.8,
  pFeature = 1,
  clusterAlg = "kmeans",
  innerLinkage = "average",
  distance = "pearson",
  changeCDFArea = 0.001,
  nbreaks = 20,
  seed = 123456789.12345,
  weightsItem = NULL,
  weightsFeature = NULL,
  corUse = "everything",
  verbose = F
)
}
\arguments{
\item{d}{Optional. A matrix where columns=items/samples and rows are features. 
For example, a gene expression matrix of genes in rows and microarrays in columns.
OR ExpressionSet object. (Default = NULL)}

\item{maxK}{Optional. An  integer value. maximum cluster number to evaluate.
(Default = 100)}

\item{reps}{Optional. An integer value. number of subsamples.  (Default = 100)}

\item{pItem}{Optional. A numerical value. proportion of items to sample.
(Default = 0.8)}

\item{pFeature}{Optional. A numerical value. proportion of features to sample.
(Default = 1)}

\item{clusterAlg}{Optional. A character value. cluster algorithm. "hc" 
heirarchical (hclust) or "km" for kmeans. (Default = "kmeans")}

\item{innerLinkage}{Optional. A heirarchical linkage method for subsampling.
(Default = "average")}

\item{distance}{Optional. A character value. sample distance measures: 
"pearson","spearman", or "euclidean". (Default = "pearson")}

\item{changeCDFArea}{Optional. Minimum spline distance for seq(2,`maxK`,
length.out = `nbreaks`) (Default = 0.001)}

\item{nbreaks}{Optional. Number of breaks to use in 
seq(2,`maxK`,length.out = `nbreaks`) this becomes the kGrid argument in 
run.consensus.cluster. (Default = 20)}

\item{seed}{Optional, A numerical value. Sets random seed for reproducible results.
(Default = 123456789.12345)}

\item{weightsItem}{Optional. A numerical vector. weights to be used for 
sampling items. (Default = NULL)}

\item{weightsFeature}{Optional. AN umerical vector. weights to be used for 
sampling features. (Default = NULL)}

\item{corUse}{Optional. Use all cores avaiable. (Default = "Everything")}

\item{verbose}{Optional. A boolean when set to TRUE, prints messages to the 
screen to indicate progress. This is useful for large datasets.(Default = FALSE)}
}
\value{
Final clustered modules.
}
\description{
A modiefed parallel version of code imported from 
https://github.com/Bioconductor-mirror/ConsensusClusterPlus 1.11.1
}
