% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.fast_greedy.once.R
\name{findModules.fast_greedy.once}
\alias{findModules.fast_greedy.once}
\title{Find Modules Fast and Greedy}
\usage{
findModules.fast_greedy.once(adj, min.module.size)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function finds modules with igraph::cluster_fast_greedy()
}
