% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.fast_greedy.R
\name{findModules.fast_greedy}
\alias{findModules.fast_greedy}
\title{Find Modules with iGraph Fast and Greedy algorithm}
\usage{
findModules.fast_greedy(adj, nperm = 10, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function wraps permutations of finding modules with 
igraph::cluster_fast_greedy().
}
