% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.igraphWrapper.R
\name{findModules.igraphWrapper}
\alias{findModules.igraphWrapper}
\title{Find Modules using igraph}
\usage{
findModules.igraphWrapper(g, method, min.module.size = 30, n_cores = 1, ...)
}
\arguments{
\item{g}{An \code{igraph} graph}

\item{method}{Which method to use to find modules. Current options supported
by this wrapper are: fast_greedy, infomap, label_prop, leading_eigen,
linkcommunities, louvain, spinglass, walktrap}

\item{min.module.size}{Optional. How many genes need to be in a module for it
to be considered valid.}

\item{n_cores}{Optional. For "megena" only, how many cores to use when
computing in parallel. If n_cores = 1, megena will not compute in parallel.}

\item{...}{Optional. Additional arguments to pass through to the cluster
algorithm.}
}
\value{
a named vector where names are genes and values are cluster assignments
}
\description{
This function finds modules with one of igraph's cluster functions, as
specified in the arguments.
}
\details{
TODO set random seed
}
\references{
Fast greedy method: http://arxiv.org/abs/cond-mat/0408187

Infomap algorithm: http://arxiv.org/abs/physics/0512106

Label propagation algorithm: Raghavan, U.N. and Albert, R. and Kumara, S.: Near linear time algorithm to detect community structures in large-scale networks. Phys Rev E 76, 036106. (2007)

Leading eigen algorithm: MEJ Newman: Finding community structure using the eigenvectors of matrices, Physical Review E 74 036104, 2006

Link communities: TODO

Louvain method: Vincent D. Blondel, Jean-Loup Guillaume, Renaud Lambiotte, Etienne Lefebvre: Fast unfolding of communities in large networks. J. Stat. Mech. (2008) P10008

Spinglass algorithm: http://arxiv.org/abs/cond-mat/0603718

Walktrap algorithm: http://arxiv.org/abs/physics/0512106
}
