% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.leading_eigen.R
\name{findModules.leading_eigen}
\alias{findModules.leading_eigen}
\title{Find Modules with Leading Eigen Edges}
\usage{
findModules.leading_eigen(adj, nperm = 10, min.module.size = 30)
}
\arguments{
\item{adj}{A n x n upper triangular adjacency in the matrix class format.}

\item{nperm}{Optional. Number of permutation on the gene ordering. (Default = 10)}

\item{min.module.size}{Optional. Integer between 1 and n genes. (Default = 30)}
}
\value{
GeneModules = n x 3 dimensional data frame with column names as Gene.ID,
moduleNumber, and moduleLabel.
}
\description{
This function iteratively wraps a function that a tries to find densely 
connected subgraphs in a graph by calculating the leading non-negative 
eigenvector of the modularity matrix of the graph.
}
