% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherEnrichment.R
\name{fisherEnrichment}
\alias{fisherEnrichment}
\title{Fishers Enrichment Analysis}
\usage{
fisherEnrichment(genesInModule, genesInGeneSet, genesInBackground)
}
\arguments{
\item{genesInModule}{Required. A character vector of differentially expressed genes or genes in a given module to test}

\item{genesInGeneSet}{Required. A character vector of genes in a gene set like GO annotations, pathways etc.}

\item{genesInBackground}{Required. A character vector of background genes}
}
\value{
A dataframe consisting of column value of fisher test P-Value. number 
of genes in the test set, enrichment set, the overlap, odds ratio and the genes
from the test set which are present in the enrichment set.
}
\description{
This function to perform Fishers enrichment analysis. Tests the enrichment of
 `genesInGeneSet` in `genesInModule` considering a bacckground gene set of 
 `genesInBackground`.
}
