% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie3.R
\name{genie3}
\alias{genie3}
\title{Run genie3}
\usage{
genie3(y, x)
}
\arguments{
\item{y}{Required. A response vector. If a factor, classification is assumed,
otherwise regression is assumed. If omitted, randomForest will run in 
unsupervised mode.}

\item{x}{Required. A data frame or a matrix of predictors, or a formula 
describing the model to be fitted (for the print method, an randomForest object).}
}
\value{
A matrix of importance measure, one row for each predictor variable. 
The column(s) are different importance measures.
}
\description{
This function runs genie3 random forest network analysis on
}
