% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetIC.R
\name{glmnetIC}
\alias{glmnetIC}
\title{Best AIC or BIC solution for Lasso or Ridge regression}
\usage{
glmnetIC(x, y, alpha = 1, ...)
}
\arguments{
\item{x}{A design matrix where rows are samples and columns are genes. Can be
in sparse or dense matrix format.}

\item{y}{The response variable (a single gene), an nx1 matrix where n = the
number of rows in x (the number of samples).}

\item{alpha}{Optional. If alpha = 1, glmnet will run lasso regression. If
alpha = 0, it will run ridge regression. If alpha > 0 and < 1, the
regression will be a mixture between the two.}

\item{...}{Optional. Other arguments accepted by \link[glmnet:glmnet]{glmnet::glmnet}}
}
\value{
a 2 x n_genes matrix, where row 1 is the coefficients for the
solution with the best AIC, and row 2 is for the best BIC. The rows are
named "AIC" and "BIC".
}
\description{
This function runs a glmnet() function with lasso or ridge regression and
returns the results with the best AIC or BIC. This function is called by both
\link{lassoIC} and \link{ridgeIC}.
}
\details{
For gaussian distributions,
\verb{AIC = 2k + n*ln(sigma^2)} and
\code{BIC = ln(n)*k + n*ln(sigma^2)},
where \code{k} is the number of parameters, \code{n} is the number of
observations, and \code{sigma^2 = sum[(y - residuals)^2] / n}.
\code{ln(sigma^2)} represents the maximum log-likelihood estimate.

See https://en.wikipedia.org/wiki/Normal_distribution#Log-likelihood and
https://en.wikipedia.org/wiki/Akaike_information_criterion#Comparison_with_least_squares
}
