% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAracne.R
\name{installAracne}
\alias{installAracne}
\title{This function applies ARACNE on the data}
\usage{
installAracne(path = NULL)
}
\arguments{
\item{path}{Optional. Path is currently not a functional argument. This function
automatically installs ARACNe from `inst/tools`. Default = NULL.}
}
\description{
This function installs the ARACNE system Function
}
