% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAracne.R
\name{installAracne}
\alias{installAracne}
\title{This function applies ARACNE on the data}
\usage{
installAracne(path = NULL, tool_storage_loc)
}
\arguments{
\item{path}{Optional. Path is currently not a functional argument. This function
automatically installs ARACNe from `inst/aracne`. Default = NULL.}

\item{tool_storage_loc}{Required. Provides the directory inside docker to 
temporarily store the ARACNE files and package. (Default = config$input_profile$temp_storage_loc)}
}
\description{
This function installs the ARACNE system Function
}
