% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModules.igraphWrapper.R
\name{linkcommunities_wrapper}
\alias{linkcommunities_wrapper}
\title{Linkcommunities wrapper}
\usage{
linkcommunities_wrapper(g, ...)
}
\arguments{
\item{g}{An \code{igraph} graph}

\item{...}{Optional. Other arguments accepted by
\code{linkcomm::linkcommunities}.}
}
\value{
a named vector where the names are gene names and the values are
  module membership.
}
\description{
This function wraps \code{linkcomm::linkcommunities}.
}
\details{
\code{linkcomm::linkcommunities} requires an edge list instead of a graph,
and returns a custom object, so this function creates the edge list and then
extracts the cluster membership from the returned object.
}
