\name{metanetwork-package}
\alias{metanetwork-package}
\alias{metanetwork}
\docType{package}
\title{metanetwork: Statistical network exploration toolkit}
\description{Package to run multiple different statistical network inference methodologies for ultra-high dimensional data and extract topological network features.}
\details{
\tabular{ll}{
Package: \tab metanetwork\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-20\cr
License: \tab GPLv3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ben Logsdon
Maintainer: Ben Logsdon <ben.logsdon@sagebase.org>
~~ The author and/or maintainer of the package ~~
}
\references{
references?
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
