% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metanetwork.R
\name{metanetwork}
\alias{metanetwork}
\title{Function to plot metanetworks using hierarchical edge bundling}
\usage{
metanetwork(
  nodes,
  links,
  colNode = viridis::viridis,
  nodeSize = 0.5,
  colLink = "#876b40",
  linkWidth = 1,
  textSize = 1,
  rad1 = 0.925,
  rad2 = 1,
  focus = NULL,
  shadowNode = TRUE,
  shadowLink = "#f4f4f4",
  export = TRUE,
  filename = "metanetwork.png",
  res = 300,
  img_size = 200,
  legend = TRUE
)
}
\arguments{
\item{nodes}{data.frame or path to csv of nodes with columns network (name of network), subnetwork (groups within networks), category (categories within groups), col (hexadecimal color of nodes, optional), size (size of nodes, optional).}

\item{links}{data.frame or path to csv of links between categories of nodes with columns from (category intiating link), to (category receiving link), col (hexadecimal color of links, optional), width (width of links, optional)}

\item{colNode}{hexadecimal for node color of length = 1 or length = \verb{nrow(nodes), or color palette used to for networks, if not provided in }nodes`. All nodes within a subnetwork must have the same color.}

\item{nodeSize}{numeric, size of nodes, if not provided in \code{nodes}. Can be a single numeric value or a numeric vector of length = \code{nrow(nodes)}. If using frequencies or other numeric values ranging from 0 to infinity, you should normalize the values so that the maximum value is at least no greater than 5 depending on the size of the graph.}

\item{colLink}{color of links, if not provided in \code{links}. Can be a single numeric value or a numeric vector of length = \code{nrow(links)}.}

\item{linkWidth}{numeric, width of links, if not provided in \code{links}. Can be a single numeric value or a numeric vector of length = \code{nrow(links)}.}

\item{textSize}{numeric, size of text displayed on figure. All text displayed is a proportion of this parameter.}

\item{rad1, rad2}{lower and upper boundary for placement of individual network titles.}

\item{focus}{character, name of network(s) to focus on; if length(focus) == 1, all links towards a single network; if length(focus) > 1, links focused on identified networks.}

\item{shadowNode}{logical, add a shadow around nodes.}

\item{shadowLink}{color of links that are not focused on.}

\item{export}{logical, export figure as png image.}

\item{filename}{filename of the (ignored if \code{export} is FALSE).}

\item{res}{resolution of png image.}

\item{img_size}{height and width of png image.}

\item{legend}{logical, add a legend to png image.}
}
\description{
Function to plot metanetworks using hierarchical edge bundling
}
\examples{
# Data 1
\dontrun{
nodes <- metanetwork:::nodes
links <- metanetwork:::links

# Figure
metanetwork(metanetwork:::nodes, metanetwork:::links)
metanetwork(metanetwork:::nodes, metanetwork:::links, focus = "Species")
metanetwork(metanetwork:::nodes, metanetwork:::links, focus = c("Species", "Drivers", "Managers"))

# Data 2
nodes <- system.file("extdata", "nodes.csv", package = "metanetwork") |> read.csv()
links <- system.file("extdata", "links.csv", package = "metanetwork") |> read.csv()
}
}
