% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrnetWrapper.R
\name{mrnetWrapper}
\alias{mrnetWrapper}
\title{Implements mrnet}
\usage{
mrnetWrapper(data, outputpath, mrnet_k = 3, debug_save = FALSE)
}
\arguments{
\item{data}{Required. A gene expression matrix with rows as sample IDs and
columns as Gene or feature IDs. The matrix will be transposed to work with
mrnet.}

\item{outputpath}{Required. The output path to save the resulting coexpression
network}

\item{mrnet_k}{Optional. The number of clusters (k) to use with `knnmi.all`.
Default is 3, which is also the default for `knnmi.all`.}

\item{debug_save}{Optional. Whether to save the network matrix as an RDS
file, for debugging purposes.}
}
\value{
a vector of file names corresponding to the files generated by this
function. The files will be `<outputpath>/aracneNetwork.csv` and
`<outputpath>/mrnetNetwork.csv`
}
\description{
This function infers the interaction network of `data` using the MRNET algorithm.
It also infers the network using ARACNE from the same package and returns both
networks.
}
