% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrnetWrapper.R
\name{mrnetWrapper}
\alias{mrnetWrapper}
\title{Implements mrnet}
\usage{
mrnetWrapper(data, path = NULL, pval = 1, outputpath)
}
\arguments{
\item{data}{Required. A gene expression matrix with rows as sample IDs and
columns as Gene or feature IDs.}

\item{path}{Optional. String containing the path to the aracne compiled 
executable. (Default = NULL)}

\item{pval}{Optional. Cutoff p-value to determine a coexpressed edge. If one is
specified aracne will produce the file `aracneNetwork.csv` if less than 1 it 
will produce the file `aracneThresholdNetwork.csv`. (Default = 1)}

\item{outputpath}{Required. The output path to save the resulting coexpression
network}
}
\value{
NULL. Saves a sparrow network object to paste0(`outputpath`,
`regressionFunction`,'mrnetNetwork.csv')
}
\description{
This function infers the interaction network of `data` using the MRNET algorithm.
}
