% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelNetworkWrapper.R
\name{parallelNetworkWrapper}
\alias{parallelNetworkWrapper}
\title{Parallel Network Wrapper}
\usage{
parallelNetworkWrapper(
  data,
  regressionFunction,
  n_cores = 1,
  cluster_type = "FORK",
  regulatorIndex = NULL,
  ...
)
}
\arguments{
\item{data}{Required. Expression matrix to be used for network construction,
which should have samples as rows and genes as columns.}

\item{regressionFunction}{Required. The name of the regression function to use,
which should exactly match the name of a function in this package. Current
options are: lassoIC, lassoCV, ridgeIC, ridgeCV, sparrowZ}

\item{n_cores}{Optional. The number of parallel cores to use.}

\item{cluster_type}{Optional. Use "FORK" if running on a Unix system, and
"PSOCK" if running on Windows.}

\item{regulatorIndex}{Optional. A vector of numerical indexes into
`colnames(data)` for a subset of genes that should be used in the network,
rather than using the full gene set.}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Wrapper to run network detection in a parallel fashion
}
