% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelNetworkWrapper.R
\name{parallelNetworkWrapper}
\alias{parallelNetworkWrapper}
\title{Parallel Network Wrapper}
\usage{
parallelNetworkWrapper(
  data,
  regressionFunction,
  log_file_path = ".",
  n_cores = 1,
  cluster_type = "FORK",
  regulatorIndex = NULL,
  ...
)
}
\arguments{
\item{data}{Expression matrix to be used for network construction, which
should have samples as rows and genes as columns.}

\item{regressionFunction}{The name of the regression function to use. Current
options are: lassoIC, lassoCV, ridgeIC, ridgeCV, vbsr}

\item{log_file_path}{Optional. The folder path to where log files should be
stored. Log files capture any output during parallel execution. If omitted,
log files will be stored in the working directory.}

\item{n_cores}{Optional. The number of parallel cores to use.}

\item{cluster_type}{Optional. Use "FORK" if running on a Unix system, and
"PSOCK" if running on Windows.}

\item{regulatorIndex}{Optional. A vector of numerical indexes into
`colnames(data)` for a subset of genes that should be used in the network.
All genes in `data` will be tested against this subset, rather than against
every gene.}

\item{...}{Optional. Additional arguments that are passed through to the
network algorithm.}
}
\value{
TODO
}
\description{
Wrapper to run network detection in a parallel fashion
}
