% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankConsensus.R
\name{rankConsensus}
\alias{rankConsensus}
\title{Ranks Consensus networks}
\usage{
rankConsensus(networks)
}
\arguments{
\item{networks}{Required. A list object containing an individual network as
a list entry.}
}
\value{
The best rank consensus network.
}
\description{
This function ranks a list of consensus network objects and returns the best
rank network. TODO
}
