% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedEdgeList.R
\name{rankedEdgeList}
\alias{rankedEdgeList}
\title{Rank edges of a coexpression matrix}
\usage{
rankedEdgeList(network, symmetric = FALSE, maxLength = 1e+07)
}
\arguments{
\item{network}{A network object as a matrix object}

\item{symmetric}{Optional. A logical `TRUE` or `FALSE` to treat `network` as
a symmetrical matrix. (Default = FALSE)}

\item{maxLength}{Optional. The maximum length of an output ranked edgelist.
(Default = 1e7)}
}
\value{
A greatest to least ranked list of eddge strenge in the form of a 
data.frame() object.
}
\description{
This functions produces a ranked edgelist of a coexpression matrix
}
