% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.consensus.cluster.R
\name{run.consensus.cluster}
\alias{run.consensus.cluster}
\title{Runs Consensus Clustering Algorithm}
\usage{
run.consensus.cluster(
  d,
  kGrid = 2:5,
  repCount = 10,
  pGenes = 0.8,
  hclust_method = "average",
  distance = "pearson",
  clusterAlg = "hclust",
  verbose = FALSE,
  corUse = "everything"
)
}
\arguments{
\item{d}{A matrix where rows are samples and columns are genes}

\item{kGrid}{Optional. A vector of k values to use for clustering.}

\item{repCount}{Optional. Replicate count for random sampling.}

\item{pGenes}{Optional. Proportion of genes to sample on each rep}

\item{hclust_method}{Optional. The "method" argument of \code{hclust}}

\item{distance}{Optional. Sample distance metrics: "pearson", "spearman", or "euclidean".}

\item{clusterAlg}{Optional. One of three cluster algorithms: "hclust",
"kmeans", or "pam".}

\item{verbose}{Optional. When TRUE, prints messages to the screen to indicate progress.}

\item{corUse}{Optional. The "use" argument of the \code{cor} function}
}
\value{
A named list containing "areaUnderCDF" = a numerical vector of areas,
and "consensus.matrix" = a list of consensus matrices. Both the vector and
the inner list are named with the "k" value that was used to create them.
}
\description{
A modified parallel version of code imported from
https://bioconductor.org/packages/ConsensusClusterPlus 1.70.0. Taken out
of findModules.consensusCluster.R
}
