% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.consensus.cluster.R
\name{run.consensus.cluster}
\alias{run.consensus.cluster}
\title{Runs Consensus Clustering Algorithm}
\usage{
run.consensus.cluster(
  rep_ind = 0,
  main.dist.obj,
  kGrid = 2:5,
  pGenes = 0.8,
  clusterAlg = "hclust",
  hclust_method = "average",
  distance_metric = "pearson",
  verbose = FALSE,
  seed = 101
)
}
\arguments{
\item{rep_ind}{Optional. The index of this random sub-sample, used as an ID
and in setting the random seed.}

\item{main.dist.obj}{A \code{dist} object containing distances between each
feature/gene in the original data. The object must have gene names as
labels.}

\item{kGrid}{Optional. A vector of k values to use for clustering.}

\item{pGenes}{Optional. Proportion of genes to sample at each random
sub-sample.}

\item{clusterAlg}{Optional. One of three cluster algorithms: "hclust",
"kmeans", or "pam".}

\item{hclust_method}{Optional. The "method" argument of \code{hclust}, if
\code{clusterAlg} is "hclust"}

\item{distance_metric}{Optional. The metric to use to calculate distances:
"pearson", "spearman", or "euclidean".}

\item{verbose}{Optional. When set to TRUE, prints messages to the screen to
indicate progress. When FALSE, no messages are printed.}

\item{seed}{Optional. The random seed to use. This value will be added to
\code{rep_ind} to create a unique seed value for each random sample, for
reproducibility.}
}
\value{
A named list where names are "k_" plus each \code{k} in \code{kGrid},
  and each item is a vector of cluster assignments generated from that
  \code{k} value. The vector's names are genes and the values are cluster
  numbers.
}
\description{
A modified parallel version of code imported from
https://bioconductor.org/packages/ConsensusClusterPlus 1.70.0. Taken out
of findModules.consensusCluster.R
}
