% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparrowZ.R
\name{sparrowZ}
\alias{sparrowZ}
\title{Runs vbsr gene across a matrix}
\usage{
sparrowZ(y, x, ...)
}
\arguments{
\item{y}{Required. response variable. Normally distributed errors for 
family="normal". For family="binomial" should be coded as a vector of 0's and 1's.}

\item{x}{Required. Design matrix, an n x m matrix, with rows as observations.}
}
\value{
A coexpression value
}
\description{
This function wraps variable bays spike regression of a genes expression across
a matrix of genes expressed in the same samples.
}
