% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigress.R
\name{tigress}
\alias{tigress}
\title{Runs tigress on an expression matrix}
\usage{
tigress(y, x)
}
\arguments{
\item{y}{Required. A vector of gene expression values.}

\item{x}{Required. A gene expression matrix.}
}
\value{
Vector of coexpression values of gene Y to columns of X
}
\description{
This function implements Trustful Inference of Gene REgulation with Stability 
Selection (TIGRESS) algoritm..
}
