% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsrWrapper.R
\name{vbsrWrapper}
\alias{vbsrWrapper}
\title{Runs variational Bayes spike regression across a gene expression matrix}
\usage{
vbsrWrapper(x, y, fdr = 0.05, ...)
}
\arguments{
\item{x}{Required. A design matrix where rows are samples and columns are
genes. Can be in sparse or dense matrix format.}

\item{y}{Required. The response variable (a single gene), an nx1 matrix where
n = the number of rows in x (the number of samples).}

\item{fdr}{Optional. FDR threshold cut off for edge determination. (Default = 0.05)}

\item{...}{Other parameters accepted by `vbsr::vbsr()`}
}
\value{
A network matrix constructed from the Z-values of the fit
}
\description{
This function wraps variable bays spike regression of a genes expression across
a matrix of genes expressed in the same samples. Returns both the original
output and the output with a 2Z cutoff.
}
