% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsrWrapper.R
\name{vbsrWrapper}
\alias{vbsrWrapper}
\title{Runs variational Bayes spike regression across a gene expression matrix}
\usage{
vbsrWrapper(x, y, fdr = 0.05, ...)
}
\arguments{
\item{x}{A design matrix where rows are samples and columns are genes. Can be
in sparse or dense matrix format.}

\item{y}{The response variable (a single gene), an nx1 matrix where n = the
number of rows in x (the number of samples).}

\item{fdr}{Optional. FDR threshold cut off for edge determination.)}

\item{...}{Other parameters accepted by \code{vbsr::vbsr()}}
}
\value{
A named list where "Z" is a network matrix constructed from the
  Z-values of the fit, and "2Z" is a network matrix where values with a
  p-value less than the FDR threshold are re-calculated. TODO
}
\description{
This function wraps variable Bayes spike regression, using a single gene's
expression across all samples as the response variable, and the expression of
the other genes as the input.
}
