% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcnaTOM.R
\name{wgcnaTOM}
\alias{wgcnaTOM}
\title{Runs WGCNA TOMsimilarity}
\usage{
wgcnaTOM(
  data,
  path = NULL,
  pval = 1,
  outputpath,
  RsquaredCut = 0.8,
  defaultNaPower = 6
)
}
\arguments{
\item{data}{Required. Gene expression data in a matrix or data frame. Rows 
correspond to samples and columns to genes.}

\item{path}{Optional. Not utilized in this function. (Default = NULL)}

\item{pval}{Optional. Not utilized in this function. (Default = 1)}

\item{outputpath}{Required. The file path to write the resulting coexpression 
network.}

\item{RsquaredCut}{Optional. Desired minimum scale free topology fitting index 
R^2. (Default = 0.80)}

\item{defaultNaPower}{Optional. The power to rais the abs(cor(data)) matrix to
if thee power estimate from WGCNA::pickSoftThreshold() is too low (Default = 6).}
}
\value{
NULL. Writes coexpression network named wgcnaTopologicalOverlapMatrixNetwork.csv
to `outputpath`
}
\description{
Runs WGCNA::TOMsimilarity over a gene expression matrix.
}
