% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorizeData.R
\name{winsorizeData}
\alias{winsorizeData}
\title{Winsorize a Gene Expression Matrix}
\usage{
winsorizeData(x)
}
\arguments{
\item{x}{Gene expression matrix}
}
\value{
A winsorized expression matrix
}
\description{
Will winsorize a given gene expression matrix
}
